% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{abs}
\alias{abs}
\alias{abs,Column-method}
\title{abs}
\usage{
\S4method{abs}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Computes the absolute value.
}
\note{
abs since 1.5.0
}
\examples{
\dontrun{abs(df$c)}
}
\seealso{
Other normal_funcs: \code{\link{bitwiseNOT}},
  \code{\link{coalesce}}, \code{\link{column}},
  \code{\link{expr}}, \code{\link{greatest}},
  \code{\link{ifelse}}, \code{\link{isnan}},
  \code{\link{least}}, \code{\link{lit}},
  \code{\link{nanvl}}, \code{\link{negate}},
  \code{\link{randn}}, \code{\link{rand}},
  \code{\link{struct}}, \code{\link{when}}
}
