% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{add_months}
\alias{add_months}
\alias{add_months}
\alias{add_months,Column,numeric-method}
\title{add_months}
\usage{
add_months(y, x)

\S4method{add_months}{Column,numeric}(y, x)
}
\arguments{
\item{y}{Column to compute on}

\item{x}{Number of months to add}
}
\description{
Returns the date that is numMonths after startDate.
}
\note{
add_months since 1.5.0
}
\examples{
\dontrun{add_months(df$d, 1)}
}
\seealso{
Other datetime_funcs: \code{\link{date_add}},
  \code{\link{date_format}}, \code{\link{date_sub}},
  \code{\link{datediff}}, \code{\link{dayofmonth}},
  \code{\link{dayofyear}}, \code{\link{from_unixtime}},
  \code{\link{from_utc_timestamp}}, \code{\link{hour}},
  \code{\link{last_day}}, \code{\link{minute}},
  \code{\link{months_between}}, \code{\link{month}},
  \code{\link{next_day}}, \code{\link{quarter}},
  \code{\link{second}}, \code{\link{to_date}},
  \code{\link{to_utc_timestamp}},
  \code{\link{unix_timestamp}}, \code{\link{weekofyear}},
  \code{\link{window}}, \code{\link{year}}
}
