% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generics.R, R/group.R
\docType{methods}
\name{agg}
\alias{agg}
\alias{agg,GroupedData-method}
\alias{count,GroupedData-method}
\alias{summarize}
\alias{summarize,GroupedData-method}
\title{Count}
\usage{
agg(x, ...)

summarize(x, ...)

\S4method{count}{GroupedData}(x)

\S4method{agg}{GroupedData}(x, ...)

\S4method{summarize}{GroupedData}(x, ...)
}
\arguments{
\item{x}{a GroupedData}

\item{x}{a GroupedData}
}
\value{
a DataFrame

a DataFrame
}
\description{
Count the number of rows for each group.
The resulting DataFrame will also contain the grouping columns.

Aggregates on the entire DataFrame without groups.
The resulting DataFrame will also contain the grouping columns.
}
\details{
df2 <- agg(df, <column> = <aggFunction>)
df2 <- agg(df, newColName = aggFunction(column))
}
\examples{
\dontrun{
  count(groupBy(df, "name"))
}
\dontrun{
 df2 <- agg(df, age = "sum")  # new column name will be created as 'SUM(age#0)'
 df2 <- agg(df, ageSum = sum(df$age)) # Creates a new column named ageSum
}
}

