% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\docType{methods}
\name{alias}
\alias{alias}
\alias{alias,Column-method}
\title{alias}
\usage{
\S4method{alias}{Column}(object, data)
}
\arguments{
\item{object}{Column to rename}

\item{data}{new name to use}
}
\description{
Set a new name for a column
}
\note{
alias since 1.4.0
}
\seealso{
Other colum_func: \code{\link{between}},
  \code{\link{cast}}, \code{\link{endsWith}},
  \code{\link{otherwise}}, \code{\link{over}},
  \code{\link{startsWith}}, \code{\link{substr}}
}
