% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllib_recommendation.R
\docType{class}
\name{ALSModel-class}
\alias{ALSModel-class}
\title{S4 class that represents an ALSModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala ALSWrapper}
}
\description{
S4 class that represents an ALSModel
}
\note{
ALSModel since 2.1.0
}
