% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{approxCountDistinct}
\alias{approxCountDistinct}
\alias{approxCountDistinct}
\alias{approxCountDistinct,Column-method}
\alias{approxCountDistinct}
\alias{approxCountDistinct,Column-method}
\title{Returns the approximate number of distinct items in a group}
\usage{
approxCountDistinct(x, ...)

\S4method{approxCountDistinct}{Column}(x, rsd = 0.05)

\S4method{approxCountDistinct}{Column}(x, rsd = 0.05)
}
\arguments{
\item{x}{Column to compute on.}

\item{...}{further arguments to be passed to or from other methods.}

\item{rsd}{maximum estimation error allowed (default = 0.05)}
}
\value{
the approximate number of distinct items in a group.
}
\description{
Returns the approximate number of distinct items in a group. This is a column
aggregate function.
}
\note{
approxCountDistinct(Column) since 1.4.0

approxCountDistinct(Column, numeric) since 1.4.0
}
\examples{
\dontrun{approxCountDistinct(df$c)}
\dontrun{approxCountDistinct(df$c, 0.02)}
}
