% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{array_contains}
\alias{array_contains}
\alias{array_contains}
\alias{array_contains,Column-method}
\title{array_contains}
\usage{
array_contains(x, value)

\S4method{array_contains}{Column}(x, value)
}
\arguments{
\item{x}{A Column}

\item{value}{A value to be checked if contained in the column}
}
\description{
Returns true if the array contain the value.
}
\note{
array_contains since 1.6.0
}
\examples{
\dontrun{array_contains(df$c, 1)}
}
\seealso{
Other collection_funcs: \code{\link{explode}},
  \code{\link{posexplode}}, \code{\link{size}},
  \code{\link{sort_array}}
}
