% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{atan2}
\alias{atan2}
\alias{atan2,Column-method}
\title{atan2}
\usage{
\S4method{atan2}{Column}(y, x)
}
\arguments{
\item{y}{Column to compute on.}

\item{x}{Column to compute on.}
}
\description{
Returns the angle theta from the conversion of rectangular coordinates (x, y) to
polar coordinates (r, theta).
}
\note{
atan2 since 1.5.0
}
\examples{
\dontrun{atan2(df$c, x)}
}
\seealso{
Other math_funcs: \code{\link{acos}}, \code{\link{asin}},
  \code{\link{atan}}, \code{\link{bin}},
  \code{\link{bround}}, \code{\link{cbrt}},
  \code{\link{ceil}}, \code{\link{conv}},
  \code{\link{corr}}, \code{\link{cosh}},
  \code{\link{cos}}, \code{\link{covar_pop}},
  \code{\link{cov}}, \code{\link{expm1}},
  \code{\link{exp}}, \code{\link{factorial}},
  \code{\link{floor}}, \code{\link{hex}},
  \code{\link{hypot}}, \code{\link{log10}},
  \code{\link{log1p}}, \code{\link{log2}},
  \code{\link{log}}, \code{\link{pmod}},
  \code{\link{rint}}, \code{\link{round}},
  \code{\link{shiftLeft}},
  \code{\link{shiftRightUnsigned}},
  \code{\link{shiftRight}}, \code{\link{signum}},
  \code{\link{sinh}}, \code{\link{sin}},
  \code{\link{sqrt}}, \code{\link{tanh}},
  \code{\link{tan}}, \code{\link{toDegrees}},
  \code{\link{toRadians}}, \code{\link{unhex}}
}
