% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{avg}
\alias{avg}
\alias{avg}
\alias{avg,Column-method}
\title{avg}
\usage{
avg(x, ...)

\S4method{avg}{Column}(x)
}
\arguments{
\item{x}{Column to compute on or a GroupedData object.}

\item{...}{additional argument(s) when \code{x} is a GroupedData object.}
}
\description{
Aggregate function: returns the average of the values in a group.
}
\note{
avg since 1.4.0
}
\examples{
\dontrun{avg(df$c)}
}
\seealso{
Other agg_funcs: \code{\link{agg}},
  \code{\link{countDistinct}}, \code{\link{count}},
  \code{\link{first}}, \code{\link{kurtosis}},
  \code{\link{last}}, \code{\link{max}},
  \code{\link{mean}}, \code{\link{min}}, \code{\link{sd}},
  \code{\link{skewness}}, \code{\link{stddev_pop}},
  \code{\link{stddev_samp}}, \code{\link{sumDistinct}},
  \code{\link{sum}}, \code{\link{var_pop}},
  \code{\link{var_samp}}, \code{\link{var}}
}
