% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{base64}
\alias{base64}
\alias{base64}
\alias{base64,Column-method}
\title{base64}
\usage{
base64(x)

\S4method{base64}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Computes the BASE64 encoding of a binary column and returns it as a string column.
This is the reverse of unbase64.
}
\note{
base64 since 1.5.0
}
\examples{
\dontrun{base64(df$c)}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{concat_ws}}, \code{\link{concat}},
  \code{\link{decode}}, \code{\link{encode}},
  \code{\link{format_number}}, \code{\link{format_string}},
  \code{\link{initcap}}, \code{\link{instr}},
  \code{\link{length}}, \code{\link{levenshtein}},
  \code{\link{locate}}, \code{\link{lower}},
  \code{\link{lpad}}, \code{\link{ltrim}},
  \code{\link{regexp_extract}},
  \code{\link{regexp_replace}}, \code{\link{reverse}},
  \code{\link{rpad}}, \code{\link{rtrim}},
  \code{\link{soundex}}, \code{\link{substring_index}},
  \code{\link{translate}}, \code{\link{trim}},
  \code{\link{unbase64}}, \code{\link{upper}}
}
