% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/column.R
\name{between}
\alias{between}
\alias{between,Column-method}
\title{between}
\usage{
between(x, bounds)

\S4method{between}{Column}(x, bounds)
}
\arguments{
\item{x}{a Column}

\item{bounds}{lower and upper bounds}
}
\description{
Test if the column is between the lower bound and upper bound, inclusive.
}
\note{
between since 1.5.0
}
\seealso{
Other column_func: 
\code{\link{alias}()},
\code{\link{cast}()},
\code{\link{endsWith}()},
\code{\link{otherwise}()},
\code{\link{over}()},
\code{\link{startsWith}()},
\code{\link{substr}()}
}
\concept{column_func}
