% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{bitwiseNOT}
\alias{bitwiseNOT}
\alias{bitwiseNOT}
\alias{bitwiseNOT,Column-method}
\title{bitwiseNOT}
\usage{
bitwiseNOT(x)

\S4method{bitwiseNOT}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Computes bitwise NOT.
}
\note{
bitwiseNOT since 1.5.0
}
\examples{
\dontrun{bitwiseNOT(df$c)}
}
\seealso{
Other normal_funcs: \code{\link{abs}},
  \code{\link{coalesce}}, \code{\link{column}},
  \code{\link{expr}}, \code{\link{greatest}},
  \code{\link{ifelse}}, \code{\link{isnan}},
  \code{\link{least}}, \code{\link{lit}},
  \code{\link{nanvl}}, \code{\link{negate}},
  \code{\link{randn}}, \code{\link{rand}},
  \code{\link{struct}}, \code{\link{when}}
}
