% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/context.R
\name{broadcast}
\alias{broadcast}
\title{Broadcast a variable to all workers}
\usage{
broadcast(sc, object)
}
\arguments{
\item{sc}{Spark Context to use}

\item{object}{Object to be broadcast}
}
\description{
Broadcast a read-only variable to the cluster, returning a \code{Broadcast}
object for reading it in distributed functions.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:2, 2L)

# Large Matrix object that we want to broadcast
randomMat <- matrix(nrow=100, ncol=10, data=rnorm(1000))
randomMatBr <- broadcast(sc, randomMat)

# Use the broadcast variable inside the function
useBroadcast <- function(x) {
  sum(value(randomMatBr) * x)
}
sumRDD <- lapply(rdd, useBroadcast)
}
}

