% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{methods}
\name{cache,DataFrame-method}
\alias{cache,DataFrame-method}
\title{Cache}
\usage{
\S4method{cache}{DataFrame}(x)
}
\arguments{
\item{x}{A SparkSQL DataFrame}
}
\description{
Persist with the default storage level (MEMORY_ONLY).
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
cache(df)
}
}

