% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{cacheTable}
\alias{cacheTable}
\title{Cache Table}
\usage{
cacheTable(sqlContext, tableName)
}
\arguments{
\item{sqlContext}{SQLContext to use}

\item{tableName}{The name of the table being cached}
}
\value{
DataFrame
}
\description{
Caches the specified table in-memory.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
registerTempTable(df, "table")
cacheTable(sqlContext, "table")
}
}

