% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{cancelJobGroup}
\alias{cancelJobGroup}
\title{Cancel active jobs for the specified group}
\usage{
cancelJobGroup(groupId)
}
\arguments{
\item{groupId}{the ID of job group to be cancelled}
}
\description{
Cancel active jobs for the specified group
}
\note{
cancelJobGroup since 1.5.0
}
\examples{
\dontrun{
sparkR.session()
cancelJobGroup("myJobGroup")
}
}
