% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{clearCache}
\alias{clearCache}
\title{Clear Cache}
\usage{
clearCache(sqlContext)
}
\arguments{
\item{sqlContext}{SQLContext to use}
}
\description{
Removes all cached tables from the in-memory cache.
}
\examples{
\dontrun{
clearCache(sqlContext)
}
}

