% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{clearJobGroup}
\alias{clearJobGroup}
\alias{clearJobGroup.default}
\title{Clear current job group ID and its description}
\usage{
\method{clearJobGroup}{default}()
}
\description{
Clear current job group ID and its description
}
\note{
clearJobGroup since 1.5.0
}
\examples{
\dontrun{
sparkR.session()
clearJobGroup()
}
}
