% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osconfig.R
\docType{class}
\name{CloudObjectStorage-class}
\alias{CloudObjectStorage}
\alias{CloudObjectStorage-class}
\title{CloudObjectStorage is a class that is designed for IBM cloud object storage (COS)
It sets up the hadoop config for COS and provide the final file url.}
\description{
sparkContext:  a SparkContext object.
'
credentials:  a dictionary with the following required keys:
  endpoint
  accessKey
  secretKey
}
\details{
configurationName: string identifies the configurations that has been
set.
When using this from a IBM Spark service instance that
is configured to connect to particular Bluemix object store
instances, the values for these credentials can be obtained
by clicking on the 'insert to code' link just below a data
source.
}

