% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{coalesce,RDD,numeric-method}
\alias{coalesce,RDD}
\alias{coalesce,RDD,numeric-method}
\title{Return a new RDD that is reduced into numPartitions partitions.}
\usage{
\S4method{coalesce}{RDD,numeric}(x, numPartitions, shuffle = FALSE)
}
\arguments{
\item{x}{The RDD.}

\item{numPartitions}{Number of partitions to create.}
}
\description{
Return a new RDD that is reduced into numPartitions partitions.
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(1, 2, 3, 4, 5), 3L)
numPartitions(rdd)               # 3
numPartitions(coalesce(rdd, 1L)) # 1
}
}
\seealso{
repartition
}

