% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{methods}
\name{collect,DataFrame-method}
\alias{collect,DataFrame-method}
\title{Collects all the elements of a Spark DataFrame and coerces them into an R data.frame.}
\usage{
\S4method{collect}{DataFrame}(x, stringsAsFactors = FALSE)
}
\arguments{
\item{x}{A SparkSQL DataFrame}

\item{stringsAsFactors}{(Optional) A logical indicating whether or not string columns
should be converted to factors. FALSE by default.}
}
\description{
Collects all the elements of a Spark DataFrame and coerces them into an R data.frame.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
collected <- collect(df)
firstName <- collected[[1]]$name
}
}

