% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/column.R
\name{asc}
\alias{asc}
\alias{asc_nulls_first}
\alias{asc_nulls_last}
\alias{contains}
\alias{desc}
\alias{desc_nulls_first}
\alias{desc_nulls_last}
\alias{getField}
\alias{getItem}
\alias{isNaN}
\alias{isNull}
\alias{isNotNull}
\alias{like}
\alias{rlike}
\alias{columnfunctions}
\title{A set of operations working with SparkDataFrame columns}
\usage{
asc(x)

asc_nulls_first(x)

asc_nulls_last(x)

contains(x, ...)

desc(x)

desc_nulls_first(x)

desc_nulls_last(x)

getField(x, ...)

getItem(x, ...)

isNaN(x)

isNull(x)

isNotNull(x)

like(x, ...)

rlike(x, ...)
}
\arguments{
\item{x}{a Column object.}

\item{...}{additional argument(s).}
}
\description{
A set of operations working with SparkDataFrame columns
}
