% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{concat_ws}
\alias{concat_ws}
\alias{concat_ws}
\alias{concat_ws,character,Column-method}
\title{concat_ws}
\usage{
concat_ws(sep, x, ...)

\S4method{concat_ws}{character,Column}(sep, x, ...)
}
\arguments{
\item{sep}{separator to use.}

\item{x}{column to concatenate.}

\item{...}{other columns to concatenate.}
}
\description{
Concatenates multiple input string columns together into a single string column,
using the given separator.
}
\note{
concat_ws since 1.5.0
}
\examples{
\dontrun{concat_ws('-', df$s, df$d)}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{base64}}, \code{\link{concat}},
  \code{\link{decode}}, \code{\link{encode}},
  \code{\link{format_number}}, \code{\link{format_string}},
  \code{\link{initcap}}, \code{\link{instr}},
  \code{\link{length}}, \code{\link{levenshtein}},
  \code{\link{locate}}, \code{\link{lower}},
  \code{\link{lpad}}, \code{\link{ltrim}},
  \code{\link{regexp_extract}},
  \code{\link{regexp_replace}}, \code{\link{reverse}},
  \code{\link{rpad}}, \code{\link{rtrim}},
  \code{\link{soundex}}, \code{\link{substring_index}},
  \code{\link{translate}}, \code{\link{trim}},
  \code{\link{unbase64}}, \code{\link{upper}}
}
