% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{conv}
\alias{conv}
\alias{conv}
\alias{conv,Column,numeric,numeric-method}
\title{conv}
\usage{
conv(x, fromBase, toBase)

\S4method{conv}{Column,numeric,numeric}(x, fromBase, toBase)
}
\arguments{
\item{x}{column to convert.}

\item{fromBase}{base to convert from.}

\item{toBase}{base to convert to.}
}
\description{
Convert a number in a string column from one base to another.
}
\note{
conv since 1.5.0
}
\examples{
\dontrun{conv(df$n, 2, 16)}
}
\seealso{
Other math_funcs: \code{\link{acos}}, \code{\link{asin}},
  \code{\link{atan2}}, \code{\link{atan}},
  \code{\link{bin}}, \code{\link{bround}},
  \code{\link{cbrt}}, \code{\link{ceil}},
  \code{\link{corr}}, \code{\link{cosh}},
  \code{\link{cos}}, \code{\link{covar_pop}},
  \code{\link{cov}}, \code{\link{expm1}},
  \code{\link{exp}}, \code{\link{factorial}},
  \code{\link{floor}}, \code{\link{hex}},
  \code{\link{hypot}}, \code{\link{log10}},
  \code{\link{log1p}}, \code{\link{log2}},
  \code{\link{log}}, \code{\link{pmod}},
  \code{\link{rint}}, \code{\link{round}},
  \code{\link{shiftLeft}},
  \code{\link{shiftRightUnsigned}},
  \code{\link{shiftRight}}, \code{\link{signum}},
  \code{\link{sinh}}, \code{\link{sin}},
  \code{\link{sqrt}}, \code{\link{tanh}},
  \code{\link{tan}}, \code{\link{toDegrees}},
  \code{\link{toRadians}}, \code{\link{unhex}}
}
