% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{methods}
\name{count,DataFrame-method}
\alias{count,DataFrame-method}
\title{Count}
\usage{
\S4method{count}{DataFrame}(x)
}
\arguments{
\item{x}{A SparkSQL DataFrame}
}
\description{
Returns the number of rows in a DataFrame
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
count(df)
}
}

