% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{countDistinct}
\alias{countDistinct}
\alias{n_distinct}
\alias{countDistinct}
\alias{countDistinct,Column-method}
\alias{n_distinct}
\alias{n_distinct,Column-method}
\title{Count Distinct Values}
\usage{
countDistinct(x, ...)

n_distinct(x, ...)

\S4method{countDistinct}{Column}(x, ...)

\S4method{n_distinct}{Column}(x, ...)
}
\arguments{
\item{x}{Column to compute on}

\item{...}{other columns}
}
\value{
the number of distinct items in a group.
}
\description{
Count Distinct Values

Aggregate function: returns the number of distinct items in a group.
}
\note{
countDistinct since 1.4.0

n_distinct since 1.4.0
}
\examples{
\dontrun{countDistinct(df$c)}
\dontrun{n_distinct(df$c)}
}
\seealso{
Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{count}}, \code{\link{first}},
  \code{\link{kurtosis}}, \code{\link{last}},
  \code{\link{max}}, \code{\link{mean}}, \code{\link{min}},
  \code{\link{sd}}, \code{\link{skewness}},
  \code{\link{stddev_pop}}, \code{\link{stddev_samp}},
  \code{\link{sumDistinct}}, \code{\link{sum}},
  \code{\link{var_pop}}, \code{\link{var_samp}},
  \code{\link{var}}
}
