% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{covar_pop}
\alias{covar_pop}
\alias{covar_pop}
\alias{covar_pop,characterOrColumn,characterOrColumn-method}
\title{covar_pop}
\usage{
covar_pop(col1, col2)

\S4method{covar_pop}{characterOrColumn,characterOrColumn}(col1, col2)
}
\arguments{
\item{col1}{First column to compute cov_pop.}

\item{col2}{Second column to compute cov_pop.}
}
\description{
Compute the population covariance between two expressions.
}
\note{
covar_pop since 2.0.0
}
\examples{
\dontrun{
covar_pop(df$c, df$d)
covar_pop("c", "d")
}
}
\seealso{
Other math_funcs: \code{\link{acos}}, \code{\link{asin}},
  \code{\link{atan2}}, \code{\link{atan}},
  \code{\link{bin}}, \code{\link{bround}},
  \code{\link{cbrt}}, \code{\link{ceil}},
  \code{\link{conv}}, \code{\link{corr}},
  \code{\link{cosh}}, \code{\link{cos}}, \code{\link{cov}},
  \code{\link{expm1}}, \code{\link{exp}},
  \code{\link{factorial}}, \code{\link{floor}},
  \code{\link{hex}}, \code{\link{hypot}},
  \code{\link{log10}}, \code{\link{log1p}},
  \code{\link{log2}}, \code{\link{log}},
  \code{\link{pmod}}, \code{\link{rint}},
  \code{\link{round}}, \code{\link{shiftLeft}},
  \code{\link{shiftRightUnsigned}},
  \code{\link{shiftRight}}, \code{\link{signum}},
  \code{\link{sinh}}, \code{\link{sin}},
  \code{\link{sqrt}}, \code{\link{tanh}},
  \code{\link{tan}}, \code{\link{toDegrees}},
  \code{\link{toRadians}}, \code{\link{unhex}}
}
