% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{crc32}
\alias{crc32}
\alias{crc32}
\alias{crc32,Column-method}
\title{crc32}
\usage{
crc32(x)

\S4method{crc32}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Calculates the cyclic redundancy check value  (CRC32) of a binary column and
returns the value as a bigint.
}
\note{
crc32 since 1.5.0
}
\examples{
\dontrun{crc32(df$c)}
}
\seealso{
Other misc_funcs: \code{\link{hash}}, \code{\link{md5}},
  \code{\link{monotonically_increasing_id}},
  \code{\link{sha1}}, \code{\link{sha2}}
}
