% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{createDataFrame}
\alias{createDataFrame}
\title{Create a DataFrame from an RDD}
\usage{
createDataFrame(sqlContext, data, schema = NULL, samplingRatio = 1)
}
\arguments{
\item{sqlContext}{A SQLContext}

\item{data}{An RDD or list or data.frame}

\item{schema}{a list of column names or named list (StructType), optional}
}
\value{
an DataFrame
}
\description{
Converts an RDD to a DataFrame by infer the types.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
rdd <- lapply(parallelize(sc, 1:10), function(x) list(a=x, b=as.character(x)))
df <- createDataFrame(sqlContext, rdd)
}
}

