% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{createExternalTable}
\alias{createExternalTable}
\title{(Deprecated) Create an external table}
\usage{
createExternalTable(tableName, path = NULL, source = NULL, schema = NULL, ...)
}
\arguments{
\item{tableName}{a name of the table.}

\item{path}{the path of files to load.}

\item{source}{the name of external data source.}

\item{schema}{the schema of the data required for some data sources.}

\item{...}{additional argument(s) passed to the method.}
}
\value{
A SparkDataFrame.
}
\description{
Creates an external table based on the dataset in a data source,
Returns a SparkDataFrame associated with the external table.
}
\details{
The data source is specified by the \code{source} and a set of options(...).
If \code{source} is not specified, the default data source configured by
"spark.sql.sources.default" will be used.
}
\note{
createExternalTable since 1.4.0
}
\examples{
\dontrun{
sparkR.session()
df <- createExternalTable("myjson", path="path/to/json", source="json", schema)
}
}
\seealso{
\link{createTable}
}
