% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{currentDatabase}
\alias{currentDatabase}
\title{Returns the current default database}
\usage{
currentDatabase()
}
\value{
name of the current default database.
}
\description{
Returns the current default database.
}
\note{
since 2.2.0
}
\examples{
\dontrun{
sparkR.session()
currentDatabase()
}
}
