% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R, R/group.R
\docType{class}
\name{DataFrame-class}
\alias{DataFrame-class}
\alias{agg}
\alias{agg,DataFrame-method}
\alias{dataFrame}
\alias{groupBy}
\alias{groupedData}
\alias{summarize}
\alias{summarize,DataFrame-method}
\title{S4 class that represents a DataFrame}
\usage{
dataFrame(sdf, isCached = FALSE)

\S4method{agg}{DataFrame}(x, ...)

\S4method{summarize}{DataFrame}(x, ...)

groupBy(x, ...)

groupedData(sgd)
}
\arguments{
\item{sdf}{A Java object reference to the backing Scala DataFrame}

\item{isCached}{TRUE if the dataFrame is cached}

\item{x}{a DataFrame}

\item{env}{An R environment that stores bookkeeping states of the DataFrame}

\item{sdf}{A Java object reference to the backing Scala DataFrame}
}
\description{
DataFrames can be created using functions like
             \code{jsonFile}, \code{table} etc.

Compute aggregates by specifying a list of columns
}
\seealso{
jsonFile, table
}

