% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{date_sub}
\alias{date_sub}
\alias{date_sub}
\alias{date_sub,Column,numeric-method}
\title{date_sub}
\usage{
date_sub(y, x)

\S4method{date_sub}{Column,numeric}(y, x)
}
\arguments{
\item{y}{Column to compute on}

\item{x}{Number of days to substract}
}
\description{
Returns the date that is \code{x} days before
}
\note{
date_sub since 1.5.0
}
\examples{
\dontrun{date_sub(df$d, 1)}
}
\seealso{
Other datetime_funcs: \code{\link{add_months}},
  \code{\link{date_add}}, \code{\link{date_format}},
  \code{\link{datediff}}, \code{\link{dayofmonth}},
  \code{\link{dayofyear}}, \code{\link{from_unixtime}},
  \code{\link{from_utc_timestamp}}, \code{\link{hour}},
  \code{\link{last_day}}, \code{\link{minute}},
  \code{\link{months_between}}, \code{\link{month}},
  \code{\link{next_day}}, \code{\link{quarter}},
  \code{\link{second}}, \code{\link{to_date}},
  \code{\link{to_utc_timestamp}},
  \code{\link{unix_timestamp}}, \code{\link{weekofyear}},
  \code{\link{window}}, \code{\link{year}}
}
