% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{datediff}
\alias{datediff}
\alias{datediff}
\alias{datediff,Column-method}
\title{datediff}
\usage{
datediff(y, x)

\S4method{datediff}{Column}(y, x)
}
\arguments{
\item{y}{end Column to use.}

\item{x}{start Column to use.}
}
\description{
Returns the number of days from \code{start} to \code{end}.
}
\note{
datediff since 1.5.0
}
\examples{
\dontrun{datediff(df$c, x)}
}
\seealso{
Other datetime_funcs: \code{\link{add_months}},
  \code{\link{date_add}}, \code{\link{date_format}},
  \code{\link{date_sub}}, \code{\link{dayofmonth}},
  \code{\link{dayofyear}}, \code{\link{from_unixtime}},
  \code{\link{from_utc_timestamp}}, \code{\link{hour}},
  \code{\link{last_day}}, \code{\link{minute}},
  \code{\link{months_between}}, \code{\link{month}},
  \code{\link{next_day}}, \code{\link{quarter}},
  \code{\link{second}}, \code{\link{to_date}},
  \code{\link{to_utc_timestamp}},
  \code{\link{unix_timestamp}}, \code{\link{weekofyear}},
  \code{\link{window}}, \code{\link{year}}
}
