% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{decode}
\alias{decode}
\alias{decode}
\alias{decode,Column,character-method}
\title{decode}
\usage{
decode(x, charset)

\S4method{decode}{Column,character}(x, charset)
}
\arguments{
\item{x}{Column to compute on.}

\item{charset}{Character set to use}
}
\description{
Computes the first argument into a string from a binary using the provided character set
(one of 'US-ASCII', 'ISO-8859-1', 'UTF-8', 'UTF-16BE', 'UTF-16LE', 'UTF-16').
}
\note{
decode since 1.6.0
}
\examples{
\dontrun{decode(df$c, "UTF-8")}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{base64}}, \code{\link{concat_ws}},
  \code{\link{concat}}, \code{\link{encode}},
  \code{\link{format_number}}, \code{\link{format_string}},
  \code{\link{initcap}}, \code{\link{instr}},
  \code{\link{length}}, \code{\link{levenshtein}},
  \code{\link{locate}}, \code{\link{lower}},
  \code{\link{lpad}}, \code{\link{ltrim}},
  \code{\link{regexp_extract}},
  \code{\link{regexp_replace}}, \code{\link{reverse}},
  \code{\link{rpad}}, \code{\link{rtrim}},
  \code{\link{soundex}}, \code{\link{substring_index}},
  \code{\link{translate}}, \code{\link{trim}},
  \code{\link{unbase64}}, \code{\link{upper}}
}
