% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{dropTempTable}
\alias{dropTempTable}
\title{(Deprecated) Drop Temporary Table}
\usage{
dropTempTable(tableName)
}
\arguments{
\item{tableName}{The name of the SparkSQL table to be dropped.}
}
\description{
Drops the temporary table with the given table name in the catalog.
If the table has been cached/persisted before, it's also unpersisted.
}
\note{
dropTempTable since 1.4.0
}
\examples{
\dontrun{
sparkR.session()
df <- read.df(path, "parquet")
createOrReplaceTempView(df, "table")
dropTempTable("table")
}
}
\seealso{
\link{dropTempView}
}
