% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{dropTempTable}
\alias{dropTempTable}
\title{Drop Temporary Table}
\usage{
dropTempTable(sqlContext, tableName)
}
\arguments{
\item{sqlContext}{SQLContext to use}

\item{tableName}{The name of the SparkSQL table to be dropped.}
}
\description{
Drops the temporary table with the given table name in the catalog.
If the table has been cached/persisted before, it's also unpersisted.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
df <- read.df(sqlContext, path, "parquet")
registerTempTable(df, "table")
dropTempTable(sqlContext, "table")
}
}

