% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{methods}
\name{dtypes,DataFrame-method}
\alias{dtypes,DataFrame-method}
\title{DataTypes}
\usage{
\S4method{dtypes}{DataFrame}(x)
}
\arguments{
\item{x}{A SparkSQL DataFrame}
}
\description{
Return all column names and their data types as a list
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
dtypes(df)
}
}

