% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/column.R
\name{endsWith}
\alias{endsWith}
\alias{endsWith,Column-method}
\title{endsWith}
\usage{
endsWith(x, suffix)

\S4method{endsWith}{Column}(x, suffix)
}
\arguments{
\item{x}{vector of character string whose "ends" are considered}

\item{suffix}{character vector (often of length one)}
}
\description{
Determines if entries of x end with string (entries of) suffix respectively,
where strings are recycled to common lengths.
}
\note{
endsWith since 1.4.0
}
\seealso{
Other column_func: 
\code{\link{alias}()},
\code{\link{between}()},
\code{\link{cast}()},
\code{\link{otherwise}()},
\code{\link{over}()},
\code{\link{startsWith}()},
\code{\link{substr}()}
}
\concept{column_func}
