% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{except,DataFrame,DataFrame-method}
\alias{except}
\alias{except,DataFrame,DataFrame-method}
\title{except}
\usage{
\S4method{except}{DataFrame,DataFrame}(x, y)

except(x, y)
}
\arguments{
\item{x}{A Spark DataFrame}

\item{y}{A Spark DataFrame}
}
\value{
A DataFrame containing the result of the except operation.
}
\description{
Return a new DataFrame containing rows in this DataFrame
but not in another DataFrame. This is equivalent to `EXCEPT` in SQL.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
df1 <- jsonFile(sqlContext, path)
df2 <- jsonFile(sqlContext, path2)
exceptDF <- except(df, df2)
}
}

