% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{explain,DataFrame-method}
\alias{explain}
\alias{explain,DataFrame-method}
\title{Explain}
\usage{
\S4method{explain}{DataFrame}(x, extended = FALSE)

explain(x, ...)
}
\arguments{
\item{x}{A SparkSQL DataFrame}

\item{extended}{Logical. If extended is False, explain() only prints the physical plan.}
}
\description{
Print the logical and physical Catalyst plans to the console for debugging.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
explain(df, TRUE)
}
}

