% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{explode}
\alias{explode}
\alias{explode}
\alias{explode,Column-method}
\title{explode}
\usage{
explode(x)

\S4method{explode}{Column}(x)
}
\arguments{
\item{x}{Column to compute on}
}
\description{
Creates a new row for each element in the given array or map column.
}
\note{
explode since 1.5.0
}
\examples{
\dontrun{explode(df$c)}
}
\seealso{
Other collection_funcs: \code{\link{array_contains}},
  \code{\link{posexplode}}, \code{\link{size}},
  \code{\link{sort_array}}
}
