% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{expr}
\alias{expr}
\alias{expr}
\alias{expr,character-method}
\title{expr}
\usage{
expr(x)

\S4method{expr}{character}(x)
}
\arguments{
\item{x}{an expression character object to be parsed.}
}
\description{
Parses the expression string into the column that it represents, similar to
SparkDataFrame.selectExpr
}
\note{
expr since 1.5.0
}
\examples{
\dontrun{expr('length(name)')}
}
\seealso{
Other normal_funcs: \code{\link{abs}},
  \code{\link{bitwiseNOT}}, \code{\link{coalesce}},
  \code{\link{column}}, \code{\link{greatest}},
  \code{\link{ifelse}}, \code{\link{isnan}},
  \code{\link{least}}, \code{\link{lit}},
  \code{\link{nanvl}}, \code{\link{negate}},
  \code{\link{randn}}, \code{\link{rand}},
  \code{\link{struct}}, \code{\link{when}}
}
