% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllib_classification.R
\docType{class}
\name{FMClassificationModel-class}
\alias{FMClassificationModel-class}
\title{S4 class that represents a FMClassificationModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala FMClassifierWrapper}
}
\description{
S4 class that represents a FMClassificationModel
}
\note{
FMClassificationModel since 3.1.0
}
