% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllib_regression.R
\docType{class}
\name{FMRegressionModel-class}
\alias{FMRegressionModel-class}
\title{S4 class that represents a FMRegressionModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala FMRegressorWrapper}
}
\description{
S4 class that represents a FMRegressionModel
}
\note{
FMRegressionModel since 3.1.0
}
