% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{format_string}
\alias{format_string}
\alias{format_string}
\alias{format_string,character,Column-method}
\title{format_string}
\usage{
format_string(format, x, ...)

\S4method{format_string}{character,Column}(format, x, ...)
}
\arguments{
\item{format}{a character object of format strings.}

\item{x}{a Column.}

\item{...}{additional Column(s).}
}
\description{
Formats the arguments in printf-style and returns the result as a string column.
}
\note{
format_string since 1.5.0
}
\examples{
\dontrun{format_string('\%d \%s', df$a, df$b)}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{base64}}, \code{\link{concat_ws}},
  \code{\link{concat}}, \code{\link{decode}},
  \code{\link{encode}}, \code{\link{format_number}},
  \code{\link{initcap}}, \code{\link{instr}},
  \code{\link{length}}, \code{\link{levenshtein}},
  \code{\link{locate}}, \code{\link{lower}},
  \code{\link{lpad}}, \code{\link{ltrim}},
  \code{\link{regexp_extract}},
  \code{\link{regexp_replace}}, \code{\link{reverse}},
  \code{\link{rpad}}, \code{\link{rtrim}},
  \code{\link{soundex}}, \code{\link{substring_index}},
  \code{\link{translate}}, \code{\link{trim}},
  \code{\link{unbase64}}, \code{\link{upper}}
}
