% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllib_fpm.R
\docType{class}
\name{FPGrowthModel-class}
\alias{FPGrowthModel-class}
\title{S4 class that represents a FPGrowthModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala FPGrowthModel}
}
\description{
S4 class that represents a FPGrowthModel
}
\note{
FPGrowthModel since 2.2.0
}
