% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{from_utc_timestamp}
\alias{from_utc_timestamp}
\alias{from_utc_timestamp}
\alias{from_utc_timestamp,Column,character-method}
\title{from_utc_timestamp}
\usage{
from_utc_timestamp(y, x)

\S4method{from_utc_timestamp}{Column,character}(y, x)
}
\arguments{
\item{y}{Column to compute on.}

\item{x}{time zone to use.}
}
\description{
Given a timestamp, which corresponds to a certain time of day in UTC, returns another timestamp
that corresponds to the same time of day in the given timezone.
}
\note{
from_utc_timestamp since 1.5.0
}
\examples{
\dontrun{from_utc_timestamp(df$t, 'PST')}
}
\seealso{
Other datetime_funcs: \code{\link{add_months}},
  \code{\link{date_add}}, \code{\link{date_format}},
  \code{\link{date_sub}}, \code{\link{datediff}},
  \code{\link{dayofmonth}}, \code{\link{dayofyear}},
  \code{\link{from_unixtime}}, \code{\link{hour}},
  \code{\link{last_day}}, \code{\link{minute}},
  \code{\link{months_between}}, \code{\link{month}},
  \code{\link{next_day}}, \code{\link{quarter}},
  \code{\link{second}}, \code{\link{to_date}},
  \code{\link{to_utc_timestamp}},
  \code{\link{unix_timestamp}}, \code{\link{weekofyear}},
  \code{\link{window}}, \code{\link{year}}
}
