% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataFrame.R
\name{generateAliasesForIntersectedCols}
\alias{generateAliasesForIntersectedCols}
\title{Creates a list of columns by replacing the intersected ones with aliases}
\usage{
generateAliasesForIntersectedCols(x, intersectedColNames, suffix)
}
\arguments{
\item{x}{a SparkDataFrame}

\item{intersectedColNames}{a list of intersected column names of the SparkDataFrame}

\item{suffix}{a suffix for the column name}
}
\value{
list of columns
}
\description{
Creates a list of columns by replacing the intersected ones with aliases.
The name of the alias column is formed by concatanating the original column name and a suffix.
}
\note{
generateAliasesForIntersectedCols since 1.6.0
}
