% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{getLocalProperty}
\alias{getLocalProperty}
\title{Get a local property set in this thread, or \code{NULL} if it is missing. See
\code{setLocalProperty}.}
\usage{
getLocalProperty(key)
}
\arguments{
\item{key}{The key for a local property.}
}
\description{
Get a local property set in this thread, or \code{NULL} if it is missing. See
\code{setLocalProperty}.
}
\note{
getLocalProperty since 2.3.0
}
\examples{
\dontrun{
getLocalProperty("spark.scheduler.pool")
}
}
