% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{greatest}
\alias{greatest}
\alias{greatest}
\alias{greatest,Column-method}
\title{greatest}
\usage{
greatest(x, ...)

\S4method{greatest}{Column}(x, ...)
}
\arguments{
\item{x}{Column to compute on}

\item{...}{other columns}
}
\description{
Returns the greatest value of the list of column names, skipping null values.
This function takes at least 2 parameters. It will return null if all parameters are null.
}
\note{
greatest since 1.5.0
}
\examples{
\dontrun{greatest(df$c, df$d)}
}
\seealso{
Other normal_funcs: \code{\link{abs}},
  \code{\link{bitwiseNOT}}, \code{\link{coalesce}},
  \code{\link{column}}, \code{\link{expr}},
  \code{\link{ifelse}}, \code{\link{isnan}},
  \code{\link{least}}, \code{\link{lit}},
  \code{\link{nanvl}}, \code{\link{negate}},
  \code{\link{randn}}, \code{\link{rand}},
  \code{\link{struct}}, \code{\link{when}}
}
