% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{groupBy,DataFrame-method}
\alias{groupBy,DataFrame-method}
\alias{group_by}
\alias{group_by,DataFrame-method}
\title{GroupBy}
\usage{
\S4method{groupBy}{DataFrame}(x, ...)

\S4method{group_by}{DataFrame}(x, ...)

group_by(x, ...)
}
\arguments{
\item{x}{a DataFrame}
}
\value{
a GroupedData
}
\description{
Groups the DataFrame using the specified columns, so we can run aggregation on them.
}
\examples{
\dontrun{
  # Compute the average for all numeric columns grouped by department.
  avg(groupBy(df, "department"))

  # Compute the max age and average salary, grouped by department and gender.
  agg(groupBy(df, "department", "gender"), salary="avg", "age" -> "max")
}
}
\seealso{
GroupedData
}

