% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/group.R
\docType{class}
\name{GroupedData-class}
\alias{GroupedData-class}
\title{S4 class that represents a GroupedData}
\arguments{
\item{sgd}{A Java object reference to the backing Scala GroupedData}
}
\description{
GroupedDatas can be created using groupBy() on a DataFrame
}
\seealso{
groupBy
}

