% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{hash}
\alias{hash}
\alias{hash}
\alias{hash,Column-method}
\title{hash}
\usage{
hash(x, ...)

\S4method{hash}{Column}(x, ...)
}
\arguments{
\item{x}{Column to compute on.}

\item{...}{additional Column(s) to be included.}
}
\description{
Calculates the hash code of given columns, and returns the result as a int column.
}
\note{
hash since 2.0.0
}
\examples{
\dontrun{hash(df$c)}
}
\seealso{
Other misc_funcs: \code{\link{crc32}}, \code{\link{md5}},
  \code{\link{monotonically_increasing_id}},
  \code{\link{sha1}}, \code{\link{sha2}}
}
