% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{methods}
\name{head,DataFrame-method}
\alias{head,DataFrame-method}
\title{Head}
\usage{
\S4method{head}{DataFrame}(x, num = 6L)
}
\arguments{
\item{x}{A SparkSQL DataFrame}

\item{num}{The number of rows to return. Default is 6.}
}
\value{
A data.frame
}
\description{
Return the first NUM rows of a DataFrame as a data.frame. If NUM is NULL,
then head() returns the first 6 rows in keeping with the current data.frame
convention in R.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
head(df)
}
}

