% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{hour}
\alias{hour}
\alias{hour}
\alias{hour,Column-method}
\title{hour}
\usage{
hour(x)

\S4method{hour}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Extracts the hours as an integer from a given date/timestamp/string.
}
\note{
hour since 1.5.0
}
\examples{
\dontrun{hour(df$c)}
}
\seealso{
Other datetime_funcs: \code{\link{add_months}},
  \code{\link{date_add}}, \code{\link{date_format}},
  \code{\link{date_sub}}, \code{\link{datediff}},
  \code{\link{dayofmonth}}, \code{\link{dayofyear}},
  \code{\link{from_unixtime}},
  \code{\link{from_utc_timestamp}}, \code{\link{last_day}},
  \code{\link{minute}}, \code{\link{months_between}},
  \code{\link{month}}, \code{\link{next_day}},
  \code{\link{quarter}}, \code{\link{second}},
  \code{\link{to_date}}, \code{\link{to_utc_timestamp}},
  \code{\link{unix_timestamp}}, \code{\link{weekofyear}},
  \code{\link{window}}, \code{\link{year}}
}
