% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{initcap}
\alias{initcap}
\alias{initcap}
\alias{initcap,Column-method}
\title{initcap}
\usage{
initcap(x)

\S4method{initcap}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Returns a new string column by converting the first letter of each word to uppercase.
Words are delimited by whitespace.
}
\details{
For example, "hello world" will become "Hello World".
}
\note{
initcap since 1.5.0
}
\examples{
\dontrun{initcap(df$c)}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{base64}}, \code{\link{concat_ws}},
  \code{\link{concat}}, \code{\link{decode}},
  \code{\link{encode}}, \code{\link{format_number}},
  \code{\link{format_string}}, \code{\link{instr}},
  \code{\link{length}}, \code{\link{levenshtein}},
  \code{\link{locate}}, \code{\link{lower}},
  \code{\link{lpad}}, \code{\link{ltrim}},
  \code{\link{regexp_extract}},
  \code{\link{regexp_replace}}, \code{\link{reverse}},
  \code{\link{rpad}}, \code{\link{rtrim}},
  \code{\link{soundex}}, \code{\link{substring_index}},
  \code{\link{translate}}, \code{\link{trim}},
  \code{\link{unbase64}}, \code{\link{upper}}
}
