% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{insertInto,DataFrame,character-method}
\alias{insertInto}
\alias{insertInto,DataFrame,character-method}
\title{insertInto}
\usage{
\S4method{insertInto}{DataFrame,character}(x, tableName, overwrite = FALSE)

insertInto(x, tableName, ...)
}
\arguments{
\item{x}{A SparkSQL DataFrame}

\item{tableName}{A character vector containing the name of the table}

\item{overwrite}{A logical argument indicating whether or not to overwrite
the existing rows in the table.}
}
\description{
Insert the contents of a DataFrame into a table registered in the current SQL Context.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
df <- read.df(sqlContext, path, "parquet")
df2 <- read.df(sqlContext, path2, "parquet")
registerTempTable(df, "table1")
insertInto(df2, "table1", overwrite = TRUE)
}
}

