% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{instr}
\alias{instr}
\alias{instr}
\alias{instr,Column,character-method}
\title{instr}
\usage{
instr(y, x)

\S4method{instr}{Column,character}(y, x)
}
\arguments{
\item{y}{column to check}

\item{x}{substring to check}
}
\description{
Locate the position of the first occurrence of substr column in the given string.
Returns null if either of the arguments are null.
}
\details{
Note: The position is not zero based, but 1 based index. Returns 0 if substr
could not be found in str.
}
\note{
instr since 1.5.0
}
\examples{
\dontrun{instr(df$c, 'b')}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{base64}}, \code{\link{concat_ws}},
  \code{\link{concat}}, \code{\link{decode}},
  \code{\link{encode}}, \code{\link{format_number}},
  \code{\link{format_string}}, \code{\link{initcap}},
  \code{\link{length}}, \code{\link{levenshtein}},
  \code{\link{locate}}, \code{\link{lower}},
  \code{\link{lpad}}, \code{\link{ltrim}},
  \code{\link{regexp_extract}},
  \code{\link{regexp_replace}}, \code{\link{reverse}},
  \code{\link{rpad}}, \code{\link{rtrim}},
  \code{\link{soundex}}, \code{\link{substring_index}},
  \code{\link{translate}}, \code{\link{trim}},
  \code{\link{unbase64}}, \code{\link{upper}}
}
