% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{intersect,DataFrame,DataFrame-method}
\alias{intersect}
\alias{intersect,DataFrame,DataFrame-method}
\title{Intersect}
\usage{
\S4method{intersect}{DataFrame,DataFrame}(x, y)

intersect(x, y)
}
\arguments{
\item{x}{A Spark DataFrame}

\item{y}{A Spark DataFrame}
}
\value{
A DataFrame containing the result of the intersect.
}
\description{
Return a new DataFrame containing rows only in both this DataFrame
and another DataFrame. This is equivalent to `INTERSECT` in SQL.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
df1 <- jsonFile(sqlContext, path)
df2 <- jsonFile(sqlContext, path2)
intersectDF <- intersect(df, df2)
}
}

