% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{isnan}
\alias{isnan}
\alias{is.nan}
\alias{is.nan,Column-method}
\alias{isnan}
\alias{isnan,Column-method}
\title{is.nan}
\usage{
isnan(x)

\S4method{is.nan}{Column}(x)

\S4method{isnan}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Return true if the column is NaN, alias for \link{isnan}
}
\note{
is.nan since 2.0.0

isnan since 2.0.0
}
\examples{
\dontrun{
is.nan(df$c)
isnan(df$c)
}
}
\seealso{
Other normal_funcs: \code{\link{abs}},
  \code{\link{bitwiseNOT}}, \code{\link{coalesce}},
  \code{\link{column}}, \code{\link{expr}},
  \code{\link{greatest}}, \code{\link{ifelse}},
  \code{\link{least}}, \code{\link{lit}},
  \code{\link{nanvl}}, \code{\link{negate}},
  \code{\link{randn}}, \code{\link{rand}},
  \code{\link{struct}}, \code{\link{when}}
}
