% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{isLocal,DataFrame-method}
\alias{isLocal}
\alias{isLocal,DataFrame-method}
\title{isLocal}
\usage{
\S4method{isLocal}{DataFrame}(x)

isLocal(x)
}
\arguments{
\item{x}{A SparkSQL DataFrame}
}
\description{
Returns True if the `collect` and `take` methods can be run locally
(without any Spark executors).
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
isLocal(df)
}
}

