% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{jsonFile}
\alias{jsonFile}
\title{Create a DataFrame from a JSON file.}
\usage{
jsonFile(sqlContext, path)
}
\arguments{
\item{sqlContext}{SQLContext to use}

\item{path}{Path of file to read. A vector of multiple paths is allowed.}
}
\value{
DataFrame
}
\description{
Loads a JSON file (one object per line), returning the result as a DataFrame
It goes through the entire dataset once to determine the schema.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
}
}

