% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllib_stat.R
\docType{class}
\name{KSTest-class}
\alias{KSTest-class}
\title{S4 class that represents an KSTest}
\arguments{
\item{jobj}{a Java object reference to the backing Scala KSTestWrapper}
}
\description{
S4 class that represents an KSTest
}
\note{
KSTest since 2.1.0
}
