% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{kurtosis}
\alias{kurtosis}
\alias{kurtosis}
\alias{kurtosis,Column-method}
\title{kurtosis}
\usage{
kurtosis(x)

\S4method{kurtosis}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Aggregate function: returns the kurtosis of the values in a group.
}
\note{
kurtosis since 1.6.0
}
\examples{
\dontrun{kurtosis(df$c)}
}
\seealso{
Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{countDistinct}}, \code{\link{count}},
  \code{\link{first}}, \code{\link{last}},
  \code{\link{max}}, \code{\link{mean}}, \code{\link{min}},
  \code{\link{sd}}, \code{\link{skewness}},
  \code{\link{stddev_pop}}, \code{\link{stddev_samp}},
  \code{\link{sumDistinct}}, \code{\link{sum}},
  \code{\link{var_pop}}, \code{\link{var_samp}},
  \code{\link{var}}
}
