% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/RDD.R
\docType{methods}
\name{lapply,DataFrame,function-method}
\alias{lapply}
\alias{lapply,DataFrame,function-method}
\alias{lapply,RDD,function-method}
\title{Apply a function to all elements}
\usage{
\S4method{lapply}{DataFrame,`function`}(X, FUN)

\S4method{lapply}{RDD,`function`}(X, FUN)
}
\arguments{
\item{X}{The RDD to apply the transformation.}

\item{FUN}{the transformation to apply on each element}
}
\value{
a new RDD created by the transformation.
}
\description{
This function creates a new RDD by applying the given transformation to all
elements of the given RDD
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10)
multiplyByTwo <- lapply(rdd, function(x) { x * 2 })
collect(multiplyByTwo) # 2,4,6...
}
}

