% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{last_day}
\alias{last_day}
\alias{last_day}
\alias{last_day,Column-method}
\title{last_day}
\usage{
last_day(x)

\S4method{last_day}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Given a date column, returns the last day of the month which the given date belongs to.
For example, input "2015-07-27" returns "2015-07-31" since July 31 is the last day of the
month in July 2015.
}
\note{
last_day since 1.5.0
}
\examples{
\dontrun{last_day(df$c)}
}
\seealso{
Other datetime_funcs: \code{\link{add_months}},
  \code{\link{date_add}}, \code{\link{date_format}},
  \code{\link{date_sub}}, \code{\link{datediff}},
  \code{\link{dayofmonth}}, \code{\link{dayofyear}},
  \code{\link{from_unixtime}},
  \code{\link{from_utc_timestamp}}, \code{\link{hour}},
  \code{\link{minute}}, \code{\link{months_between}},
  \code{\link{month}}, \code{\link{next_day}},
  \code{\link{quarter}}, \code{\link{second}},
  \code{\link{to_date}}, \code{\link{to_utc_timestamp}},
  \code{\link{unix_timestamp}}, \code{\link{weekofyear}},
  \code{\link{window}}, \code{\link{year}}
}
