% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{least}
\alias{least}
\alias{least}
\alias{least,Column-method}
\title{least}
\usage{
least(x, ...)

\S4method{least}{Column}(x, ...)
}
\arguments{
\item{x}{Column to compute on}

\item{...}{other columns}
}
\description{
Returns the least value of the list of column names, skipping null values.
This function takes at least 2 parameters. It will return null if all parameters are null.
}
\note{
least since 1.5.0
}
\examples{
\dontrun{least(df$c, df$d)}
}
\seealso{
Other normal_funcs: \code{\link{abs}},
  \code{\link{bitwiseNOT}}, \code{\link{coalesce}},
  \code{\link{column}}, \code{\link{expr}},
  \code{\link{greatest}}, \code{\link{ifelse}},
  \code{\link{isnan}}, \code{\link{lit}},
  \code{\link{nanvl}}, \code{\link{negate}},
  \code{\link{randn}}, \code{\link{rand}},
  \code{\link{struct}}, \code{\link{when}}
}
