% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{levenshtein}
\alias{levenshtein}
\alias{levenshtein}
\alias{levenshtein,Column-method}
\title{levenshtein}
\usage{
levenshtein(y, x)

\S4method{levenshtein}{Column}(y, x)
}
\arguments{
\item{y}{Column to compute on.}

\item{x}{Column to compute on.}
}
\description{
Computes the Levenshtein distance of the two given string columns.
}
\note{
levenshtein since 1.5.0
}
\examples{
\dontrun{levenshtein(df$c, x)}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{base64}}, \code{\link{concat_ws}},
  \code{\link{concat}}, \code{\link{decode}},
  \code{\link{encode}}, \code{\link{format_number}},
  \code{\link{format_string}}, \code{\link{initcap}},
  \code{\link{instr}}, \code{\link{length}},
  \code{\link{locate}}, \code{\link{lower}},
  \code{\link{lpad}}, \code{\link{ltrim}},
  \code{\link{regexp_extract}},
  \code{\link{regexp_replace}}, \code{\link{reverse}},
  \code{\link{rpad}}, \code{\link{rtrim}},
  \code{\link{soundex}}, \code{\link{substring_index}},
  \code{\link{translate}}, \code{\link{trim}},
  \code{\link{unbase64}}, \code{\link{upper}}
}
