% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{limit,DataFrame,numeric-method}
\alias{limit}
\alias{limit,DataFrame,numeric-method}
\title{Limit}
\usage{
\S4method{limit}{DataFrame,numeric}(x, num)

limit(x, num)
}
\arguments{
\item{x}{A SparkSQL DataFrame}

\item{num}{The number of rows to return}
}
\value{
A new DataFrame containing the number of rows specified.
}
\description{
Limit the resulting DataFrame to the number of rows specified.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
limitedDF <- limit(df, 10)
}
}

