% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllib_regression.R
\docType{class}
\name{LinearRegressionModel-class}
\alias{LinearRegressionModel-class}
\title{S4 class that represents a LinearRegressionModel}
\arguments{
\item{jobj}{a Java object reference to the backing Scala LinearRegressionWrapper}
}
\description{
S4 class that represents a LinearRegressionModel
}
\note{
LinearRegressionModel since 3.1.0
}
