% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{listColumns}
\alias{listColumns}
\title{Returns a list of columns for the given table/view in the specified database}
\usage{
listColumns(tableName, databaseName = NULL)
}
\arguments{
\item{tableName}{the qualified or unqualified name that designates a table/view. If no database
identifier is provided, it refers to a table/view in the current database.
If \code{databaseName} parameter is specified, this must be an unqualified name.}

\item{databaseName}{(optional) name of the database}
}
\value{
a SparkDataFrame of the list of column descriptions.
}
\description{
Returns a list of columns for the given table/view in the specified database.
}
\note{
since 2.2.0
}
\examples{
\dontrun{
sparkR.session()
listColumns("mytable")
}
}
