% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{listDatabases}
\alias{listDatabases}
\title{Returns a list of databases available}
\usage{
listDatabases()
}
\value{
a SparkDataFrame of the list of databases.
}
\description{
Returns a list of databases available.
}
\note{
since 2.2.0
}
\examples{
\dontrun{
sparkR.session()
listDatabases()
}
}
