% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{listFunctions}
\alias{listFunctions}
\title{Returns a list of functions registered in the specified database}
\usage{
listFunctions(databaseName = NULL)
}
\arguments{
\item{databaseName}{(optional) name of the database}
}
\value{
a SparkDataFrame of the list of function descriptions.
}
\description{
Returns a list of functions registered in the specified database.
This includes all temporary functions.
}
\note{
since 2.2.0
}
\examples{
\dontrun{
sparkR.session()
listFunctions()
}
}
