% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{listTables}
\alias{listTables}
\title{Returns a list of tables or views in the specified database}
\usage{
listTables(databaseName = NULL)
}
\arguments{
\item{databaseName}{(optional) name of the database}
}
\value{
a SparkDataFrame of the list of tables.
}
\description{
Returns a list of tables or views in the specified database.
This includes all temporary views.
}
\note{
since 2.2.0
}
\examples{
\dontrun{
sparkR.session()
listTables()
listTables("default")
}
}
\seealso{
\link{tables}
}
