% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{lit}
\alias{lit}
\alias{lit}
\alias{lit,ANY-method}
\title{lit}
\usage{
lit(x)

\S4method{lit}{ANY}(x)
}
\arguments{
\item{x}{a literal value or a Column.}
}
\description{
A new \linkS4class{Column} is created to represent the literal value.
If the parameter is a \linkS4class{Column}, it is returned unchanged.
}
\note{
lit since 1.5.0
}
\examples{
\dontrun{
lit(df$name)
select(df, lit("x"))
select(df, lit("2015-01-01"))
}
}
\seealso{
Other normal_funcs: \code{\link{abs}},
  \code{\link{bitwiseNOT}}, \code{\link{coalesce}},
  \code{\link{column}}, \code{\link{expr}},
  \code{\link{greatest}}, \code{\link{ifelse}},
  \code{\link{isnan}}, \code{\link{least}},
  \code{\link{nanvl}}, \code{\link{negate}},
  \code{\link{randn}}, \code{\link{rand}},
  \code{\link{struct}}, \code{\link{when}}
}
