% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{locate}
\alias{locate}
\alias{locate}
\alias{locate,character,Column-method}
\title{locate}
\usage{
locate(substr, str, ...)

\S4method{locate}{character,Column}(substr, str, pos = 1)
}
\arguments{
\item{substr}{a character string to be matched.}

\item{str}{a Column where matches are sought for each entry.}

\item{...}{further arguments to be passed to or from other methods.}

\item{pos}{start position of search.}
}
\description{
Locate the position of the first occurrence of substr.
}
\details{
Note: The position is not zero based, but 1 based index. Returns 0 if substr
could not be found in str.
}
\note{
locate since 1.5.0
}
\examples{
\dontrun{locate('b', df$c, 1)}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{base64}}, \code{\link{concat_ws}},
  \code{\link{concat}}, \code{\link{decode}},
  \code{\link{encode}}, \code{\link{format_number}},
  \code{\link{format_string}}, \code{\link{initcap}},
  \code{\link{instr}}, \code{\link{length}},
  \code{\link{levenshtein}}, \code{\link{lower}},
  \code{\link{lpad}}, \code{\link{ltrim}},
  \code{\link{regexp_extract}},
  \code{\link{regexp_replace}}, \code{\link{reverse}},
  \code{\link{rpad}}, \code{\link{rtrim}},
  \code{\link{soundex}}, \code{\link{substring_index}},
  \code{\link{translate}}, \code{\link{trim}},
  \code{\link{unbase64}}, \code{\link{upper}}
}
