% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{\%in\%}
\alias{\%in\%}
\alias{\%in\%,Column-method}
\title{Match a column with given values.}
\usage{
\S4method{\%in\%}{Column}(x, table)
}
\arguments{
\item{x}{a Column.}

\item{table}{a collection of values (coercible to list) to compare with.}
}
\value{
A matched values as a result of comparing with given values.
}
\description{
Match a column with given values.
}
\note{
\%in\% since 1.5.0
}
\examples{
\dontrun{
filter(df, "age in (10, 30)")
where(df, df$age \%in\% c(10, 30))
}
}
