% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{md5}
\alias{md5}
\alias{md5}
\alias{md5,Column-method}
\title{md5}
\usage{
md5(x)

\S4method{md5}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Calculates the MD5 digest of a binary column and returns the value
as a 32 character hex string.
}
\note{
md5 since 1.5.0
}
\examples{
\dontrun{md5(df$c)}
}
\seealso{
Other misc_funcs: \code{\link{crc32}}, \code{\link{hash}},
  \code{\link{monotonically_increasing_id}},
  \code{\link{sha1}}, \code{\link{sha2}}
}
