% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{mean}
\alias{mean}
\alias{mean,Column-method}
\title{mean}
\usage{
\S4method{mean}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Aggregate function: returns the average of the values in a group.
Alias for avg.
}
\note{
mean since 1.5.0
}
\examples{
\dontrun{mean(df$c)}
}
\seealso{
Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{countDistinct}}, \code{\link{count}},
  \code{\link{first}}, \code{\link{kurtosis}},
  \code{\link{last}}, \code{\link{max}}, \code{\link{min}},
  \code{\link{sd}}, \code{\link{skewness}},
  \code{\link{stddev_pop}}, \code{\link{stddev_samp}},
  \code{\link{sumDistinct}}, \code{\link{sum}},
  \code{\link{var_pop}}, \code{\link{var_samp}},
  \code{\link{var}}
}
