% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{min}
\alias{min}
\alias{min,Column-method}
\title{min}
\usage{
\S4method{min}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Aggregate function: returns the minimum value of the expression in a group.
}
\note{
min since 1.5.0
}
\examples{
\dontrun{min(df$c)}
}
\seealso{
Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{countDistinct}}, \code{\link{count}},
  \code{\link{first}}, \code{\link{kurtosis}},
  \code{\link{last}}, \code{\link{max}},
  \code{\link{mean}}, \code{\link{sd}},
  \code{\link{skewness}}, \code{\link{stddev_pop}},
  \code{\link{stddev_samp}}, \code{\link{sumDistinct}},
  \code{\link{sum}}, \code{\link{var_pop}},
  \code{\link{var_samp}}, \code{\link{var}}
}
