% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{months_between}
\alias{months_between}
\alias{months_between}
\alias{months_between,Column-method}
\title{months_between}
\usage{
months_between(y, x)

\S4method{months_between}{Column}(y, x)
}
\arguments{
\item{y}{end Column to use.}

\item{x}{start Column to use.}
}
\description{
Returns number of months between dates \code{date1} and \code{date2}.
}
\note{
months_between since 1.5.0
}
\examples{
\dontrun{months_between(df$c, x)}
}
\seealso{
Other datetime_funcs: \code{\link{add_months}},
  \code{\link{date_add}}, \code{\link{date_format}},
  \code{\link{date_sub}}, \code{\link{datediff}},
  \code{\link{dayofmonth}}, \code{\link{dayofyear}},
  \code{\link{from_unixtime}},
  \code{\link{from_utc_timestamp}}, \code{\link{hour}},
  \code{\link{last_day}}, \code{\link{minute}},
  \code{\link{month}}, \code{\link{next_day}},
  \code{\link{quarter}}, \code{\link{second}},
  \code{\link{to_date}}, \code{\link{to_utc_timestamp}},
  \code{\link{unix_timestamp}}, \code{\link{weekofyear}},
  \code{\link{window}}, \code{\link{year}}
}
