% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{nanvl}
\alias{nanvl}
\alias{nanvl}
\alias{nanvl,Column-method}
\title{nanvl}
\usage{
nanvl(y, x)

\S4method{nanvl}{Column}(y, x)
}
\arguments{
\item{y}{second Column.}

\item{x}{first Column.}
}
\description{
Returns col1 if it is not NaN, or col2 if col1 is NaN.
Both inputs should be floating point columns (DoubleType or FloatType).
}
\note{
nanvl since 1.5.0
}
\examples{
\dontrun{nanvl(df$c, x)}
}
\seealso{
Other normal_funcs: \code{\link{abs}},
  \code{\link{bitwiseNOT}}, \code{\link{coalesce}},
  \code{\link{column}}, \code{\link{expr}},
  \code{\link{greatest}}, \code{\link{ifelse}},
  \code{\link{isnan}}, \code{\link{least}},
  \code{\link{lit}}, \code{\link{negate}},
  \code{\link{randn}}, \code{\link{rand}},
  \code{\link{struct}}, \code{\link{when}}
}
