% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{negate}
\alias{negate}
\alias{negate}
\alias{negate,Column-method}
\title{negate}
\usage{
negate(x)

\S4method{negate}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Unary minus, i.e. negate the expression.
}
\note{
negate since 1.5.0
}
\examples{
\dontrun{negate(df$c)}
}
\seealso{
Other normal_funcs: \code{\link{abs}},
  \code{\link{bitwiseNOT}}, \code{\link{coalesce}},
  \code{\link{column}}, \code{\link{expr}},
  \code{\link{greatest}}, \code{\link{ifelse}},
  \code{\link{isnan}}, \code{\link{least}},
  \code{\link{lit}}, \code{\link{nanvl}},
  \code{\link{randn}}, \code{\link{rand}},
  \code{\link{struct}}, \code{\link{when}}
}
