% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RDD.R
\docType{methods}
\name{numPartitions,RDD-method}
\alias{numPartitions,RDD-method}
\title{Gets the number of partitions of an RDD}
\usage{
\S4method{numPartitions}{RDD}(x)
}
\arguments{
\item{x}{A RDD.}
}
\value{
the number of partitions of rdd as an integer.
}
\description{
Gets the number of partitions of an RDD
}
\examples{
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, 1:10, 2L)
numPartitions(rdd)  # 2L
}
}

