% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/column.R
\name{otherwise}
\alias{otherwise}
\alias{otherwise,Column-method}
\title{otherwise}
\usage{
otherwise(x, value)

\S4method{otherwise}{Column}(x, value)
}
\arguments{
\item{x}{a Column.}

\item{value}{value to replace when the corresponding entry in \code{x} is NA.
Can be a single value or a Column.}
}
\description{
If values in the specified column are null, returns the value.
Can be used in conjunction with \code{when} to specify a default value for expressions.
}
\note{
otherwise since 1.5.0
}
\seealso{
Other column_func: 
\code{\link{alias}()},
\code{\link{between}()},
\code{\link{cast}()},
\code{\link{endsWith}()},
\code{\link{over}()},
\code{\link{startsWith}()},
\code{\link{substr}()}
}
\concept{column_func}
