% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{methods}
\name{persist,DataFrame,character-method}
\alias{persist,DataFrame,character-method}
\title{Persist}
\usage{
\S4method{persist}{DataFrame,character}(x, newLevel)
}
\arguments{
\item{x}{The DataFrame to persist}
}
\description{
Persist this DataFrame with the specified storage level. For details of the
supported storage levels, refer to
http://spark.apache.org/docs/latest/programming-guide.html#rdd-persistence.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
persist(df, "MEMORY_AND_DISK")
}
}

