% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Spark.R
\name{pi}
\alias{pi}
\title{Spark function}
\usage{
pi()
}
\arguments{
\item{none}{no parameters are required}
}
\value{
default value of 1
}
\description{
This script reads the aggregated_measurements table, searches for the latest record for each device and calculates the Asset Health Index (AHI).
}
\note{
This is a spark function
}
\examples{
pi()
}

