% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{posexplode}
\alias{posexplode}
\alias{posexplode}
\alias{posexplode,Column-method}
\title{posexplode}
\usage{
posexplode(x)

\S4method{posexplode}{Column}(x)
}
\arguments{
\item{x}{Column to compute on}
}
\description{
Creates a new row for each element with position in the given array or map column.
}
\note{
posexplode since 2.1.0
}
\examples{
\dontrun{posexplode(df$c)}
}
\seealso{
Other collection_funcs: \code{\link{array_contains}},
  \code{\link{explode}}, \code{\link{size}},
  \code{\link{sort_array}}
}
