% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllib_fpm.R
\docType{class}
\name{PrefixSpan-class}
\alias{PrefixSpan-class}
\title{S4 class that represents a PrefixSpan}
\arguments{
\item{jobj}{a Java object reference to the backing Scala PrefixSpan}
}
\description{
S4 class that represents a PrefixSpan
}
\note{
PrefixSpan since 3.0.0
}
