% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema.R
\name{print.structType}
\alias{print.structType}
\title{Print a Spark StructType.}
\usage{
\method{print}{structType}(x, ...)
}
\arguments{
\item{x}{A StructType object}

\item{...}{further arguments passed to or from other methods}
}
\description{
This function prints the contents of a StructType returned from the
SparkR JVM backend.
}

