% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{methods}
\name{printSchema,DataFrame-method}
\alias{printSchema,DataFrame-method}
\title{Print Schema of a DataFrame}
\usage{
\S4method{printSchema}{DataFrame}(x)
}
\arguments{
\item{x}{A SparkSQL DataFrame}
}
\description{
Prints out the schema in tree format
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
printSchema(df)
}
}

