% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{rand}
\alias{rand}
\alias{rand}
\alias{rand,missing-method}
\alias{rand}
\alias{rand,numeric-method}
\title{rand}
\usage{
rand(seed)

\S4method{rand}{missing}(seed)

\S4method{rand}{numeric}(seed)
}
\arguments{
\item{seed}{a random seed. Can be missing.}
}
\description{
Generate a random column with independent and identically distributed (i.i.d.) samples
from U[0.0, 1.0].
}
\note{
rand since 1.5.0

rand(numeric) since 1.5.0
}
\examples{
\dontrun{rand()}
}
\seealso{
Other normal_funcs: \code{\link{abs}},
  \code{\link{bitwiseNOT}}, \code{\link{coalesce}},
  \code{\link{column}}, \code{\link{expr}},
  \code{\link{greatest}}, \code{\link{ifelse}},
  \code{\link{isnan}}, \code{\link{least}},
  \code{\link{lit}}, \code{\link{nanvl}},
  \code{\link{negate}}, \code{\link{randn}},
  \code{\link{struct}}, \code{\link{when}}
}
