% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{randn}
\alias{randn}
\alias{randn}
\alias{randn,missing-method}
\alias{randn}
\alias{randn,numeric-method}
\title{randn}
\usage{
randn(seed)

\S4method{randn}{missing}(seed)

\S4method{randn}{numeric}(seed)
}
\arguments{
\item{seed}{a random seed. Can be missing.}
}
\description{
Generate a column with independent and identically distributed (i.i.d.) samples from
the standard normal distribution.
}
\note{
randn since 1.5.0

randn(numeric) since 1.5.0
}
\examples{
\dontrun{randn()}
}
\seealso{
Other normal_funcs: \code{\link{abs}},
  \code{\link{bitwiseNOT}}, \code{\link{coalesce}},
  \code{\link{column}}, \code{\link{expr}},
  \code{\link{greatest}}, \code{\link{ifelse}},
  \code{\link{isnan}}, \code{\link{least}},
  \code{\link{lit}}, \code{\link{nanvl}},
  \code{\link{negate}}, \code{\link{rand}},
  \code{\link{struct}}, \code{\link{when}}
}
