% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{read.df}
\alias{read.df}
\title{Load an DataFrame}
\usage{
read.df(sqlContext, path = NULL, source = NULL, ...)
}
\arguments{
\item{sqlContext}{SQLContext to use}

\item{path}{The path of files to load}

\item{source}{the name of external data source}
}
\value{
DataFrame
}
\description{
Returns the dataset in a data source as a DataFrame
}
\details{
The data source is specified by the `source` and a set of options(...).
If `source` is not specified, the default data source configured by
"spark.sql.sources.default" will be used.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
df <- read.df(sqlContext, "path/to/file.json", source = "json")
}
}

