% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mllib_utils.R
\name{read.ml}
\alias{read.ml}
\title{Load a fitted MLlib model from the input path.}
\usage{
read.ml(path)
}
\arguments{
\item{path}{path of the model to read.}
}
\value{
A fitted MLlib model.
}
\description{
Load a fitted MLlib model from the input path.
}
\note{
read.ml since 2.0.0
}
\examples{
\dontrun{
path <- "path/to/model"
model <- read.ml(path)
}
}
\seealso{
\link{write.ml}
}
