% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{read.parquet}
\alias{read.parquet}
\title{Create a SparkDataFrame from a Parquet file.}
\usage{
read.parquet(path, ...)
}
\arguments{
\item{path}{path of file to read. A vector of multiple paths is allowed.}

\item{...}{additional data source specific named properties.}
}
\value{
SparkDataFrame
}
\description{
Loads a Parquet file, returning the result as a SparkDataFrame.
}
\note{
read.parquet since 1.6.0
}
