% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{read.text}
\alias{read.text}
\alias{read.text.default}
\title{Create a SparkDataFrame from a text file.}
\usage{
\method{read.text}{default}(path, ...)
}
\arguments{
\item{path}{Path of file to read. A vector of multiple paths is allowed.}

\item{...}{additional external data source specific named properties.}
}
\value{
SparkDataFrame
}
\description{
Loads text files and returns a SparkDataFrame whose schema starts with
a string column named "value", and followed by partitioned columns if
there are any.
}
\details{
Each line in the text file is a new row in the resulting SparkDataFrame.
}
\note{
read.text since 1.6.1
}
\examples{
\dontrun{
sparkR.session()
path <- "path/to/file.txt"
df <- read.text(path)
}
}
