% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{recoverPartitions}
\alias{recoverPartitions}
\title{Recovers all the partitions in the directory of a table and update the catalog}
\usage{
recoverPartitions(tableName)
}
\arguments{
\item{tableName}{the qualified or unqualified name that designates a table. If no database
identifier is provided, it refers to a table in the current database.}
}
\description{
Recovers all the partitions in the directory of a table and update the catalog. The name should
reference a partitioned table, and not a view.
}
\note{
since 2.2.0
}
\examples{
\dontrun{
sparkR.session()
recoverPartitions("myTable")
}
}
