% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{refreshByPath}
\alias{refreshByPath}
\title{Invalidates and refreshes all the cached data and metadata for SparkDataFrame containing path}
\usage{
refreshByPath(path)
}
\arguments{
\item{path}{the path of the data source.}
}
\description{
Invalidates and refreshes all the cached data (and the associated metadata) for any
SparkDataFrame that contains the given data source path. Path matching is by prefix, i.e. "/"
would invalidate everything that is cached.
}
\note{
since 2.2.0
}
\examples{
\dontrun{
sparkR.session()
refreshByPath("/path")
}
}
