% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{regexp_replace}
\alias{regexp_replace}
\alias{regexp_replace}
\alias{regexp_replace,Column,character,character-method}
\title{regexp_replace}
\usage{
regexp_replace(x, pattern, replacement)

\S4method{regexp_replace}{Column,character,character}(x, pattern, replacement)
}
\arguments{
\item{x}{a string Column.}

\item{pattern}{a regular expression.}

\item{replacement}{a character string that a matched \code{pattern} is replaced with.}
}
\description{
Replace all substrings of the specified string value that match regexp with rep.
}
\note{
regexp_replace since 1.5.0
}
\examples{
\dontrun{regexp_replace(df$c, '(\\\\d+)', '--')}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{base64}}, \code{\link{concat_ws}},
  \code{\link{concat}}, \code{\link{decode}},
  \code{\link{encode}}, \code{\link{format_number}},
  \code{\link{format_string}}, \code{\link{initcap}},
  \code{\link{instr}}, \code{\link{length}},
  \code{\link{levenshtein}}, \code{\link{locate}},
  \code{\link{lower}}, \code{\link{lpad}},
  \code{\link{ltrim}}, \code{\link{regexp_extract}},
  \code{\link{reverse}}, \code{\link{rpad}},
  \code{\link{rtrim}}, \code{\link{soundex}},
  \code{\link{substring_index}}, \code{\link{translate}},
  \code{\link{trim}}, \code{\link{unbase64}},
  \code{\link{upper}}
}
