% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{registerTempTable,DataFrame,character-method}
\alias{registerTempTable}
\alias{registerTempTable,DataFrame,character-method}
\title{Register Temporary Table}
\usage{
\S4method{registerTempTable}{DataFrame,character}(x, tableName)

registerTempTable(x, tableName)
}
\arguments{
\item{x}{A SparkSQL DataFrame}

\item{tableName}{A character vector containing the name of the table}
}
\description{
Registers a DataFrame as a Temporary Table in the SQLContext
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
registerTempTable(df, "json_df")
new_df <- sql(sqlContext, "SELECT * FROM json_df")
}
}

