% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/RDD.R
\docType{methods}
\name{repartition,DataFrame,numeric-method}
\alias{repartition,DataFrame,numeric-method}
\alias{repartition,RDD}
\alias{repartition,RDD,numeric-method}
\title{Repartition}
\usage{
\S4method{repartition}{DataFrame,numeric}(x, numPartitions)

\S4method{repartition}{RDD,numeric}(x, numPartitions)
}
\arguments{
\item{x}{A SparkSQL DataFrame}

\item{numPartitions}{The number of partitions to use.}

\item{x}{The RDD.}

\item{numPartitions}{Number of partitions to create.}
}
\description{
Return a new DataFrame that has exactly numPartitions partitions.

Return a new RDD that has exactly numPartitions partitions.
Can increase or decrease the level of parallelism in this RDD. Internally,
this uses a shuffle to redistribute data.
If you are decreasing the number of partitions in this RDD, consider using
coalesce, which can avoid performing a shuffle.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
newDF <- repartition(df, 2L)
}
\dontrun{
sc <- sparkR.init()
rdd <- parallelize(sc, list(1, 2, 3, 4, 5, 6, 7), 4L)
numPartitions(rdd)                   # 4
numPartitions(repartition(rdd, 2L))  # 2
}
}
\seealso{
coalesce
}

