% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{rpad}
\alias{rpad}
\alias{rpad}
\alias{rpad,Column,numeric,character-method}
\title{rpad}
\usage{
rpad(x, len, pad)

\S4method{rpad}{Column,numeric,character}(x, len, pad)
}
\arguments{
\item{x}{the string Column to be right-padded.}

\item{len}{maximum length of each output result.}

\item{pad}{a character string to be padded with.}
}
\description{
Right-padded with pad to a length of len.
}
\note{
rpad since 1.5.0
}
\examples{
\dontrun{rpad(df$c, 6, '#')}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{base64}}, \code{\link{concat_ws}},
  \code{\link{concat}}, \code{\link{decode}},
  \code{\link{encode}}, \code{\link{format_number}},
  \code{\link{format_string}}, \code{\link{initcap}},
  \code{\link{instr}}, \code{\link{length}},
  \code{\link{levenshtein}}, \code{\link{locate}},
  \code{\link{lower}}, \code{\link{lpad}},
  \code{\link{ltrim}}, \code{\link{regexp_extract}},
  \code{\link{regexp_replace}}, \code{\link{reverse}},
  \code{\link{rtrim}}, \code{\link{soundex}},
  \code{\link{substring_index}}, \code{\link{translate}},
  \code{\link{trim}}, \code{\link{unbase64}},
  \code{\link{upper}}
}
