% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{sample,DataFrame,logical,numeric-method}
\alias{sample}
\alias{sample,DataFrame,logical,numeric-method}
\alias{sample_frac}
\alias{sample_frac,DataFrame,logical,numeric-method}
\title{Sample}
\usage{
\S4method{sample}{DataFrame,logical,numeric}(x, withReplacement, fraction)

\S4method{sample_frac}{DataFrame,logical,numeric}(x, withReplacement, fraction)

sample(x, withReplacement, fraction, seed)

sample_frac(x, withReplacement, fraction, seed)
}
\arguments{
\item{x}{A SparkSQL DataFrame}

\item{withReplacement}{Sampling with replacement or not}

\item{fraction}{The (rough) sample target fraction}
}
\description{
Return a sampled subset of this DataFrame using a random seed.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
collect(sample(df, FALSE, 0.5))
collect(sample(df, TRUE, 0.5))
}
}

