% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{schema,DataFrame-method}
\alias{columns}
\alias{dtypes}
\alias{printSchema}
\alias{schema}
\alias{schema,DataFrame-method}
\title{Get schema object}
\usage{
\S4method{schema}{DataFrame}(x)

columns(x)

dtypes(x)

printSchema(x)

schema(x)
}
\arguments{
\item{x}{A SparkSQL DataFrame}
}
\description{
Returns the schema of this DataFrame as a structType object.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
dfSchema <- schema(df)
}
}

