% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{sd}
\alias{sd}
\alias{stddev}
\alias{sd}
\alias{sd,Column-method}
\alias{stddev}
\alias{stddev,Column-method}
\title{sd}
\usage{
sd(x, na.rm = FALSE)

stddev(x)

\S4method{sd}{Column}(x)

\S4method{stddev}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}

\item{na.rm}{currently not used.}
}
\description{
Aggregate function: alias for \link{stddev_samp}
}
\note{
sd since 1.6.0

stddev since 1.6.0
}
\examples{
\dontrun{
stddev(df$c)
select(df, stddev(df$age))
agg(df, sd(df$age))
}
}
\seealso{
\link{stddev_pop}, \link{stddev_samp}

Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{countDistinct}}, \code{\link{count}},
  \code{\link{first}}, \code{\link{kurtosis}},
  \code{\link{last}}, \code{\link{max}},
  \code{\link{mean}}, \code{\link{min}},
  \code{\link{skewness}}, \code{\link{stddev_pop}},
  \code{\link{stddev_samp}}, \code{\link{sumDistinct}},
  \code{\link{sum}}, \code{\link{var_pop}},
  \code{\link{var_samp}}, \code{\link{var}}
}
