% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{methods}
\name{selectExpr,DataFrame,character-method}
\alias{selectExpr,DataFrame,character-method}
\title{SelectExpr}
\usage{
\S4method{selectExpr}{DataFrame,character}(x, expr, ...)
}
\arguments{
\item{x}{A DataFrame to be selected from.}

\item{expr}{A string containing a SQL expression}

\item{...}{Additional expressions}
}
\value{
A DataFrame
}
\description{
Select from a DataFrame using a set of SQL expressions.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
selectExpr(df, "col1", "(col2 * 5) as newCol")
}
}

