% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{setCheckpointDir}
\alias{setCheckpointDir}
\title{Set checkpoint directory}
\usage{
setCheckpointDir(directory)
}
\arguments{
\item{directory}{Directory path to checkpoint to}
}
\description{
Set the directory under which SparkDataFrame are going to be checkpointed. The directory must be
an HDFS path if running on a cluster.
}
\note{
setCheckpointDir since 2.2.0
}
\examples{
\dontrun{
setCheckpointDir("/checkpoint")
}
}
\seealso{
\link{checkpoint}
}
