% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{setCurrentDatabase}
\alias{setCurrentDatabase}
\title{Sets the current default database}
\usage{
setCurrentDatabase(databaseName)
}
\arguments{
\item{databaseName}{name of the database}
}
\description{
Sets the current default database.
}
\note{
since 2.2.0
}
\examples{
\dontrun{
sparkR.session()
setCurrentDatabase("default")
}
}
