% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{setJobDescription}
\alias{setJobDescription}
\title{Set a human readable description of the current job.}
\usage{
setJobDescription(value)
}
\arguments{
\item{value}{The job description of the current job.}
}
\description{
Set a description that is shown as a job description in UI.
}
\note{
setJobDescription since 2.3.0
}
\examples{
\dontrun{
setJobDescription("This is an example job.")
}
}
