% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{setJobGroup}
\alias{setJobGroup}
\title{Assigns a group ID to all the jobs started by this thread until the group ID is set to a
different value or cleared.}
\usage{
setJobGroup(groupId, description, interruptOnCancel)
}
\arguments{
\item{groupId}{the ID to be assigned to job groups.}

\item{description}{description for the job group ID.}

\item{interruptOnCancel}{flag to indicate if the job is interrupted on job cancellation.}
}
\description{
Assigns a group ID to all the jobs started by this thread until the group ID is set to a
different value or cleared.
}
\note{
setJobGroup since 1.5.0
}
\examples{
\dontrun{
sparkR.session()
setJobGroup("myJobGroup", "My job group description", TRUE)
}
}
