% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{setLocalProperty}
\alias{setLocalProperty}
\title{Set a local property that affects jobs submitted from this thread, such as the
Spark fair scheduler pool.}
\usage{
setLocalProperty(key, value)
}
\arguments{
\item{key}{The key for a local property.}

\item{value}{The value for a local property.}
}
\description{
Set a local property that affects jobs submitted from this thread, such as the
Spark fair scheduler pool.
}
\note{
setLocalProperty since 2.3.0
}
\examples{
\dontrun{
setLocalProperty("spark.scheduler.pool", "poolA")
}
}
