% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{setLogLevel}
\alias{setLogLevel}
\title{Set new log level}
\usage{
setLogLevel(level)
}
\arguments{
\item{level}{New log level}
}
\description{
Set new log level: "ALL", "DEBUG", "ERROR", "FATAL", "INFO", "OFF", "TRACE", "WARN"
}
\note{
setLogLevel since 2.0.0
}
\examples{
\dontrun{
setLogLevel("ERROR")
}
}
