% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{sha1}
\alias{sha1}
\alias{sha1}
\alias{sha1,Column-method}
\title{sha1}
\usage{
sha1(x)

\S4method{sha1}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Calculates the SHA-1 digest of a binary column and returns the value
as a 40 character hex string.
}
\note{
sha1 since 1.5.0
}
\examples{
\dontrun{sha1(df$c)}
}
\seealso{
Other misc_funcs: \code{\link{crc32}}, \code{\link{hash}},
  \code{\link{md5}},
  \code{\link{monotonically_increasing_id}},
  \code{\link{sha2}}
}
