% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{sha2}
\alias{sha2}
\alias{sha2}
\alias{sha2,Column,numeric-method}
\title{sha2}
\usage{
sha2(y, x)

\S4method{sha2}{Column,numeric}(y, x)
}
\arguments{
\item{y}{column to compute SHA-2 on.}

\item{x}{one of 224, 256, 384, or 512.}
}
\description{
Calculates the SHA-2 family of hash functions of a binary column and
returns the value as a hex string.
}
\note{
sha2 since 1.5.0
}
\examples{
\dontrun{sha2(df$c, 256)}
}
\seealso{
Other misc_funcs: \code{\link{crc32}}, \code{\link{hash}},
  \code{\link{md5}},
  \code{\link{monotonically_increasing_id}},
  \code{\link{sha1}}
}
