% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/column.R, R/group.R
\docType{methods}
\name{show,DataFrame-method}
\alias{show,Column-method}
\alias{show,DataFrame-method}
\alias{show,GroupedData-method}
\title{show}
\usage{
\S4method{show}{DataFrame}(object)

\S4method{show}{Column}(object)

\S4method{show}{GroupedData}(object)
}
\arguments{
\item{x}{A SparkSQL DataFrame}
}
\description{
Print the DataFrame column names and types
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
df
}
}

