% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{showDF,DataFrame-method}
\alias{showDF}
\alias{showDF,DataFrame-method}
\title{showDF}
\usage{
\S4method{showDF}{DataFrame}(x, numRows = 20)

showDF(x, ...)
}
\arguments{
\item{x}{A SparkSQL DataFrame}

\item{numRows}{The number of rows to print. Defaults to 20.}
}
\description{
Print the first numRows rows of a DataFrame
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
showDF(df)
}
}

