% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{methods}
\name{sin}
\alias{sin}
\alias{sin,Column-method}
\title{sin}
\usage{
\S4method{sin}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Computes the sine of the given value.
}
\note{
sin since 1.5.0
}
\examples{
\dontrun{sin(df$c)}
}
\seealso{
Other math_funcs: \code{\link{acos}}, \code{\link{asin}},
  \code{\link{atan2}}, \code{\link{atan}},
  \code{\link{bin}}, \code{\link{bround}},
  \code{\link{cbrt}}, \code{\link{ceil}},
  \code{\link{conv}}, \code{\link{corr}},
  \code{\link{cosh}}, \code{\link{cos}},
  \code{\link{covar_pop}}, \code{\link{cov}},
  \code{\link{expm1}}, \code{\link{exp}},
  \code{\link{factorial}}, \code{\link{floor}},
  \code{\link{hex}}, \code{\link{hypot}},
  \code{\link{log10}}, \code{\link{log1p}},
  \code{\link{log2}}, \code{\link{log}},
  \code{\link{pmod}}, \code{\link{rint}},
  \code{\link{round}}, \code{\link{shiftLeft}},
  \code{\link{shiftRightUnsigned}},
  \code{\link{shiftRight}}, \code{\link{signum}},
  \code{\link{sinh}}, \code{\link{sqrt}},
  \code{\link{tanh}}, \code{\link{tan}},
  \code{\link{toDegrees}}, \code{\link{toRadians}},
  \code{\link{unhex}}
}
