% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{size}
\alias{size}
\alias{size}
\alias{size,Column-method}
\title{size}
\usage{
size(x)

\S4method{size}{Column}(x)
}
\arguments{
\item{x}{Column to compute on}
}
\description{
Returns length of array or map.
}
\note{
size since 1.5.0
}
\examples{
\dontrun{size(df$c)}
}
\seealso{
Other collection_funcs: \code{\link{array_contains}},
  \code{\link{explode}}, \code{\link{posexplode}},
  \code{\link{sort_array}}
}
