% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osconfig.R
\docType{class}
\name{softlayer-class}
\alias{softlayer}
\alias{softlayer-class}
\title{sparkcontext is a SparkContext object.}
\description{
name is a string that identifies this configuration. You can
    use any string you like. This allows you to create
    multiple configurations to different Object Storage accounts.
auth_url, username and password are string credentials for your
Softlayer Object Store
}

