% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{sort_array}
\alias{sort_array}
\alias{sort_array}
\alias{sort_array,Column-method}
\title{sort_array}
\usage{
sort_array(x, asc = TRUE)

\S4method{sort_array}{Column}(x, asc = TRUE)
}
\arguments{
\item{x}{A Column to sort}

\item{asc}{A logical flag indicating the sorting order.
TRUE, sorting is in ascending order.
FALSE, sorting is in descending order.}
}
\description{
Sorts the input array in ascending or descending order according
to the natural ordering of the array elements.
}
\note{
sort_array since 1.6.0
}
\examples{
\dontrun{
sort_array(df$c)
sort_array(df$c, FALSE)
}
}
\seealso{
Other collection_funcs: \code{\link{array_contains}},
  \code{\link{explode}}, \code{\link{posexplode}},
  \code{\link{size}}
}
