% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{spark.getSparkFiles}
\alias{spark.getSparkFiles}
\title{Get the absolute path of a file added through spark.addFile.}
\usage{
spark.getSparkFiles(fileName)
}
\arguments{
\item{fileName}{The name of the file added through spark.addFile}
}
\value{
the absolute path of a file added through spark.addFile.
}
\description{
Get the absolute path of a file added through spark.addFile.
}
\note{
spark.getSparkFiles since 2.1.0
}
\examples{
\dontrun{
spark.getSparkFiles("myfile")
}
}
