% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{spark.getSparkFilesRootDirectory}
\alias{spark.getSparkFilesRootDirectory}
\title{Get the root directory that contains files added through spark.addFile.}
\usage{
spark.getSparkFilesRootDirectory()
}
\value{
the root directory that contains files added through spark.addFile
}
\description{
Get the root directory that contains files added through spark.addFile.
}
\note{
spark.getSparkFilesRootDirectory since 2.1.0
}
\examples{
\dontrun{
spark.getSparkFilesRootDirectory()
}
}
