% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{spark_partition_id}
\alias{spark_partition_id}
\alias{spark_partition_id}
\alias{spark_partition_id,missing-method}
\title{Return the partition ID as a column}
\usage{
spark_partition_id(x = "missing")

\S4method{spark_partition_id}{missing}()
}
\arguments{
\item{x}{empty. Should be used with no argument.}
}
\description{
Return the partition ID as a SparkDataFrame column.
Note that this is nondeterministic because it depends on data partitioning and
task scheduling.
}
\details{
This is equivalent to the SPARK_PARTITION_ID function in SQL.
}
\note{
spark_partition_id since 2.0.0
}
\examples{
\dontrun{select(df, spark_partition_id())}
}
