% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkr-package.R
\docType{package}
\name{sparkr-package}
\alias{sparkr-package}
\title{sparkr: Gives an idea or spark of inspiration for an activity to do}
\description{
A package for giving you a random idea of something to do when you are lacking inspiration. Ideas are predetermined in a list and randomly selected for you to enjoy. Arguments for this function are simply TRUE or FALSE.

This package is an exercise in package creation using
R studio. The package includes a sample function and
a sample dataset with their respective documentation.
}
\references{
\url{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/liamthepineapple/sparkr}
  \item Report bugs at \url{https://github.com/liamthepineapple/sparkr/issues}
}


Useful links:
\itemize{
  \item \url{https://github.com/liamthepineapple/sparkr}
  \item Report bugs at \url{https://github.com/liamthepineapple/sparkr/issues}
}

}
\author{
\strong{Maintainer}: Liam Surry \email{surryl@mcmaster.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})


Liam Surry, Department of Chemistry and Chemical Biology \email{surryl@mcmaster.ca}
}
\keyword{internal}
