% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{sparkR.session.stop}
\alias{sparkR.session.stop}
\alias{sparkR.stop}
\title{Stop the Spark Session and Spark Context}
\usage{
sparkR.session.stop()

sparkR.stop()
}
\description{
Stop the Spark Session and Spark Context.
}
\details{
Also terminates the backend this R session is connected to.
}
\note{
sparkR.session.stop since 2.0.0

sparkR.stop since 1.4.0
}
