% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{sparkR.uiWebUrl}
\alias{sparkR.uiWebUrl}
\title{Get the URL of the SparkUI instance for the current active SparkSession}
\usage{
sparkR.uiWebUrl()
}
\value{
the SparkUI URL, or NA if it is disabled, or not started.
}
\description{
Get the URL of the SparkUI instance for the current active SparkSession.
}
\note{
sparkR.uiWebUrl since 2.1.1
}
\examples{
\dontrun{
sparkR.session()
url <- sparkR.uiWebUrl()
}
}
