% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{sparkR.version}
\alias{sparkR.version}
\title{Get version of Spark on which this application is running}
\usage{
sparkR.version()
}
\value{
a character string of the Spark version
}
\description{
Get version of Spark on which this application is running.
}
\note{
sparkR.version since 2.0.1
}
\examples{
\dontrun{
sparkR.session()
version <- sparkR.version()
}
}
