% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkr_dataset.R
\docType{data}
\name{sparkr_dataset}
\alias{sparkr_dataset}
\title{A dataset containing ideas for sparks categorized into 5 categories: Physical, Creative, Social, Wellness, and Bad.}
\format{
A data frame with two columns:
\describe{
\item{category}{A character string representing the five categories of the data.}
\item{ideas}{A character string containing multiple ideas or sparks for the function.}
}
}
\source{
Generated from user input.
}
\usage{
sparkr_dataset
}
\description{
A dataset containing ideas for sparks categorized into 5 categories: Physical, Creative, Social, Wellness, and Bad.
}
\keyword{datasets}
