% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{sparkRHive.init}
\alias{sparkRHive.init}
\title{(Deprecated) Initialize a new HiveContext}
\usage{
sparkRHive.init(jsc = NULL)
}
\arguments{
\item{jsc}{The existing JavaSparkContext created with SparkR.init()}
}
\description{
This function creates a HiveContext from an existing JavaSparkContext
}
\details{
Starting SparkR 2.0, a SparkSession is initialized and returned instead.
This API is deprecated and kept for backward compatibility only.
}
\note{
sparkRHive.init since 1.4.0
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRHive.init(sc)
}
}
\seealso{
\link{sparkR.session}
}
