% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparkR.R
\name{sparkRHive.init}
\alias{sparkRHive.init}
\title{Initialize a new HiveContext.}
\usage{
sparkRHive.init(jsc = NULL)
}
\arguments{
\item{jsc}{The existing JavaSparkContext created with SparkR.init()}
}
\description{
This function creates a HiveContext from an existing JavaSparkContext
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRHive.init(sc)
}
}

