% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkR.R
\name{sparkRSQL.init}
\alias{sparkRSQL.init}
\title{(Deprecated) Initialize a new SQLContext}
\usage{
sparkRSQL.init(jsc = NULL)
}
\arguments{
\item{jsc}{The existing JavaSparkContext created with SparkR.init()}
}
\description{
This function creates a SparkContext from an existing JavaSparkContext and
then uses it to initialize a new SQLContext
}
\details{
Starting SparkR 2.0, a SparkSession is initialized and returned instead.
This API is deprecated and kept for backward compatibility only.
}
\note{
sparkRSQL.init since 1.4.0
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
}
}
\seealso{
\link{sparkR.session}
}
