% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sparkR.R
\name{sparkRSQL.init}
\alias{sparkRSQL.init}
\title{Initialize a new SQLContext.}
\usage{
sparkRSQL.init(jsc = NULL)
}
\arguments{
\item{jsc}{The existing JavaSparkContext created with SparkR.init()}
}
\description{
This function creates a SparkContext from an existing JavaSparkContext and
then uses it to initialize a new SQLContext
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
}
}

