% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{sql}
\alias{sql}
\title{SQL Query}
\usage{
sql(sqlContext, sqlQuery)
}
\arguments{
\item{sqlContext}{SQLContext to use}

\item{sqlQuery}{A character vector containing the SQL query}
}
\value{
DataFrame
}
\description{
Executes a SQL query using Spark, returning the result as a DataFrame.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
registerTempTable(df, "table")
new_df <- sql(sqlContext, "SELECT * FROM table")
}
}

