% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/column.R
\name{startsWith}
\alias{startsWith}
\alias{startsWith,Column-method}
\title{startsWith}
\usage{
startsWith(x, prefix)

\S4method{startsWith}{Column}(x, prefix)
}
\arguments{
\item{x}{vector of character string whose "starts" are considered}

\item{prefix}{character vector (often of length one)}
}
\description{
Determines if entries of x start with string (entries of) prefix respectively,
where strings are recycled to common lengths.
}
\note{
startsWith since 1.4.0
}
\seealso{
Other column_func: 
\code{\link{alias}()},
\code{\link{between}()},
\code{\link{cast}()},
\code{\link{endsWith}()},
\code{\link{otherwise}()},
\code{\link{over}()},
\code{\link{substr}()}
}
\concept{column_func}
