% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{stddev_pop}
\alias{stddev_pop}
\alias{stddev_pop}
\alias{stddev_pop,Column-method}
\title{stddev_pop}
\usage{
stddev_pop(x)

\S4method{stddev_pop}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Aggregate function: returns the population standard deviation of the expression in a group.
}
\note{
stddev_pop since 1.6.0
}
\examples{
\dontrun{stddev_pop(df$c)}
}
\seealso{
\link{sd}, \link{stddev_samp}

Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{countDistinct}}, \code{\link{count}},
  \code{\link{first}}, \code{\link{kurtosis}},
  \code{\link{last}}, \code{\link{max}},
  \code{\link{mean}}, \code{\link{min}}, \code{\link{sd}},
  \code{\link{skewness}}, \code{\link{stddev_samp}},
  \code{\link{sumDistinct}}, \code{\link{sum}},
  \code{\link{var_pop}}, \code{\link{var_samp}},
  \code{\link{var}}
}
