% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{stddev_samp}
\alias{stddev_samp}
\alias{stddev_samp}
\alias{stddev_samp,Column-method}
\title{stddev_samp}
\usage{
stddev_samp(x)

\S4method{stddev_samp}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Aggregate function: returns the unbiased sample standard deviation of the expression in a group.
}
\note{
stddev_samp since 1.6.0
}
\examples{
\dontrun{stddev_samp(df$c)}
}
\seealso{
\link{stddev_pop}, \link{sd}

Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{countDistinct}}, \code{\link{count}},
  \code{\link{first}}, \code{\link{kurtosis}},
  \code{\link{last}}, \code{\link{max}},
  \code{\link{mean}}, \code{\link{min}}, \code{\link{sd}},
  \code{\link{skewness}}, \code{\link{stddev_pop}},
  \code{\link{sumDistinct}}, \code{\link{sum}},
  \code{\link{var_pop}}, \code{\link{var_samp}},
  \code{\link{var}}
}
