% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{struct}
\alias{struct}
\alias{struct}
\alias{struct,characterOrColumn-method}
\title{struct}
\usage{
struct(x, ...)

\S4method{struct}{characterOrColumn}(x, ...)
}
\arguments{
\item{x}{a column to compute on.}

\item{...}{optional column(s) to be included.}
}
\description{
Creates a new struct column that composes multiple input columns.
}
\note{
struct since 1.6.0
}
\examples{
\dontrun{
struct(df$c, df$d)
struct("col1", "col2")
}
}
\seealso{
Other normal_funcs: \code{\link{abs}},
  \code{\link{bitwiseNOT}}, \code{\link{coalesce}},
  \code{\link{column}}, \code{\link{expr}},
  \code{\link{greatest}}, \code{\link{ifelse}},
  \code{\link{isnan}}, \code{\link{least}},
  \code{\link{lit}}, \code{\link{nanvl}},
  \code{\link{negate}}, \code{\link{randn}},
  \code{\link{rand}}, \code{\link{when}}
}
