% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema.R
\name{structField}
\alias{structField}
\title{structField}
\usage{
structField(x, ...)
}
\arguments{
\item{x}{The name of the field}

\item{type}{The data type of the field}

\item{nullable}{A logical vector indicating whether or not the field is nullable}
}
\value{
a structField object
}
\description{
Create a structField object that contains the metadata for a single field in a schema.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
rdd <- lapply(parallelize(sc, 1:10), function(x) { list(x, as.character(x)) })
field1 <- structField("a", "integer", TRUE)
field2 <- structField("b", "string", TRUE)
schema <- structType(field1, field2)
df <- createDataFrame(sqlCtx, rdd, schema)
}
}

