% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/schema.R
\name{structType}
\alias{structType}
\title{structType}
\usage{
structType(x, ...)
}
\arguments{
\item{x}{a structField object (created with the field() function)}

\item{...}{additional structField objects}
}
\value{
a structType object
}
\description{
Create a structType object that contains the metadata for a DataFrame. Intended for
use with createDataFrame and toDF.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlCtx <- sparkRSQL.init(sc)
rdd <- lapply(parallelize(sc, 1:10), function(x) { list(x, as.character(x)) })
schema <- structType(structField("a", "integer"), structField("b", "string"))
df <- createDataFrame(sqlCtx, rdd, schema)
}
}

