% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\docType{methods}
\name{substr}
\alias{substr}
\alias{substr,Column-method}
\title{substr}
\usage{
\S4method{substr}{Column}(x, start, stop)
}
\arguments{
\item{x}{a Column.}

\item{start}{starting position.}

\item{stop}{ending position.}
}
\description{
An expression that returns a substring.
}
\note{
substr since 1.4.0
}
\seealso{
Other colum_func: \code{\link{alias}},
  \code{\link{between}}, \code{\link{cast}},
  \code{\link{endsWith}}, \code{\link{otherwise}},
  \code{\link{over}}, \code{\link{startsWith}}
}
