% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{sumDistinct}
\alias{sumDistinct}
\alias{sumDistinct}
\alias{sumDistinct,Column-method}
\title{sumDistinct}
\usage{
sumDistinct(x)

\S4method{sumDistinct}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Aggregate function: returns the sum of distinct values in the expression.
}
\note{
sumDistinct since 1.4.0
}
\examples{
\dontrun{sumDistinct(df$c)}
}
\seealso{
Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{countDistinct}}, \code{\link{count}},
  \code{\link{first}}, \code{\link{kurtosis}},
  \code{\link{last}}, \code{\link{max}},
  \code{\link{mean}}, \code{\link{min}}, \code{\link{sd}},
  \code{\link{skewness}}, \code{\link{stddev_pop}},
  \code{\link{stddev_samp}}, \code{\link{sum}},
  \code{\link{var_pop}}, \code{\link{var_samp}},
  \code{\link{var}}
}
