% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{table}
\alias{table}
\title{Create a DataFrame from a SparkSQL Table}
\usage{
table(sqlContext, tableName)
}
\arguments{
\item{sqlContext}{SQLContext to use}

\item{tableName}{The SparkSQL Table to convert to a DataFrame.}
}
\value{
DataFrame
}
\description{
Returns the specified Table as a DataFrame.  The Table must have already been registered
in the SQLContext.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
registerTempTable(df, "table")
new_df <- table(sqlContext, "table")
}
}

