% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{tableNames}
\alias{tableNames}
\title{Table Names}
\usage{
tableNames(sqlContext, databaseName = NULL)
}
\arguments{
\item{sqlContext}{SQLContext to use}

\item{databaseName}{name of the database}
}
\value{
a list of table names
}
\description{
Returns the names of tables in the given database as an array.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
tableNames(sqlContext, "hive")
}
}

