% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{tables}
\alias{tables}
\title{Tables}
\usage{
tables(sqlContext, databaseName = NULL)
}
\arguments{
\item{sqlContext}{SQLContext to use}

\item{databaseName}{name of the database}
}
\value{
a DataFrame
}
\description{
Returns a DataFrame containing names of tables in the given database.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
tables(sqlContext, "hive")
}
}

