% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{methods}
\name{take,DataFrame,numeric-method}
\alias{take,DataFrame,numeric-method}
\title{Take the first NUM rows of a DataFrame and return a the results as a data.frame}
\usage{
\S4method{take}{DataFrame,numeric}(x, num)
}
\description{
Take the first NUM rows of a DataFrame and return a the results as a data.frame
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
take(df, 2)
}
}

