% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{to_date}
\alias{to_date}
\alias{to_date}
\alias{to_date,Column-method}
\title{to_date}
\usage{
to_date(x)

\S4method{to_date}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Converts the column into DateType.
}
\note{
to_date since 1.5.0
}
\examples{
\dontrun{to_date(df$c)}
}
\seealso{
Other datetime_funcs: \code{\link{add_months}},
  \code{\link{date_add}}, \code{\link{date_format}},
  \code{\link{date_sub}}, \code{\link{datediff}},
  \code{\link{dayofmonth}}, \code{\link{dayofyear}},
  \code{\link{from_unixtime}},
  \code{\link{from_utc_timestamp}}, \code{\link{hour}},
  \code{\link{last_day}}, \code{\link{minute}},
  \code{\link{months_between}}, \code{\link{month}},
  \code{\link{next_day}}, \code{\link{quarter}},
  \code{\link{second}}, \code{\link{to_utc_timestamp}},
  \code{\link{unix_timestamp}}, \code{\link{weekofyear}},
  \code{\link{window}}, \code{\link{year}}
}
