% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{to_utc_timestamp}
\alias{to_utc_timestamp}
\alias{to_utc_timestamp}
\alias{to_utc_timestamp,Column,character-method}
\title{to_utc_timestamp}
\usage{
to_utc_timestamp(y, x)

\S4method{to_utc_timestamp}{Column,character}(y, x)
}
\arguments{
\item{y}{Column to compute on}

\item{x}{timezone to use}
}
\description{
Given a timestamp, which corresponds to a certain time of day in the given timezone, returns
another timestamp that corresponds to the same time of day in UTC.
}
\note{
to_utc_timestamp since 1.5.0
}
\examples{
\dontrun{to_utc_timestamp(df$t, 'PST')}
}
\seealso{
Other datetime_funcs: \code{\link{add_months}},
  \code{\link{date_add}}, \code{\link{date_format}},
  \code{\link{date_sub}}, \code{\link{datediff}},
  \code{\link{dayofmonth}}, \code{\link{dayofyear}},
  \code{\link{from_unixtime}},
  \code{\link{from_utc_timestamp}}, \code{\link{hour}},
  \code{\link{last_day}}, \code{\link{minute}},
  \code{\link{months_between}}, \code{\link{month}},
  \code{\link{next_day}}, \code{\link{quarter}},
  \code{\link{second}}, \code{\link{to_date}},
  \code{\link{unix_timestamp}}, \code{\link{weekofyear}},
  \code{\link{window}}, \code{\link{year}}
}
