% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{toRadians}
\alias{toRadians}
\alias{toRadians}
\alias{toRadians,Column-method}
\title{toRadians}
\usage{
toRadians(x)

\S4method{toRadians}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Converts an angle measured in degrees to an approximately equivalent angle measured in radians.
}
\note{
toRadians since 1.4.0
}
\examples{
\dontrun{toRadians(df$c)}
}
\seealso{
Other math_funcs: \code{\link{acos}}, \code{\link{asin}},
  \code{\link{atan2}}, \code{\link{atan}},
  \code{\link{bin}}, \code{\link{bround}},
  \code{\link{cbrt}}, \code{\link{ceil}},
  \code{\link{conv}}, \code{\link{corr}},
  \code{\link{cosh}}, \code{\link{cos}},
  \code{\link{covar_pop}}, \code{\link{cov}},
  \code{\link{expm1}}, \code{\link{exp}},
  \code{\link{factorial}}, \code{\link{floor}},
  \code{\link{hex}}, \code{\link{hypot}},
  \code{\link{log10}}, \code{\link{log1p}},
  \code{\link{log2}}, \code{\link{log}},
  \code{\link{pmod}}, \code{\link{rint}},
  \code{\link{round}}, \code{\link{shiftLeft}},
  \code{\link{shiftRightUnsigned}},
  \code{\link{shiftRight}}, \code{\link{signum}},
  \code{\link{sinh}}, \code{\link{sin}},
  \code{\link{sqrt}}, \code{\link{tanh}},
  \code{\link{tan}}, \code{\link{toDegrees}},
  \code{\link{unhex}}
}
