% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{toRDD,DataFrame-method}
\alias{toRDD}
\alias{toRDD,DataFrame-method}
\title{toRDD()}
\usage{
\S4method{toRDD}{DataFrame}(x)

toRDD(x)
}
\arguments{
\item{x}{A Spark DataFrame}
}
\description{
Converts a Spark DataFrame to an RDD while preserving column names.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
rdd <- toRDD(df)
}
}

