% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{trim}
\alias{trim}
\alias{trim}
\alias{trim,Column-method}
\title{trim}
\usage{
trim(x)

\S4method{trim}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Trim the spaces from both ends for the specified string column.
}
\note{
trim since 1.5.0
}
\examples{
\dontrun{trim(df$c)}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{base64}}, \code{\link{concat_ws}},
  \code{\link{concat}}, \code{\link{decode}},
  \code{\link{encode}}, \code{\link{format_number}},
  \code{\link{format_string}}, \code{\link{initcap}},
  \code{\link{instr}}, \code{\link{length}},
  \code{\link{levenshtein}}, \code{\link{locate}},
  \code{\link{lower}}, \code{\link{lpad}},
  \code{\link{ltrim}}, \code{\link{regexp_extract}},
  \code{\link{regexp_replace}}, \code{\link{reverse}},
  \code{\link{rpad}}, \code{\link{rtrim}},
  \code{\link{soundex}}, \code{\link{substring_index}},
  \code{\link{translate}}, \code{\link{unbase64}},
  \code{\link{upper}}
}
