% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{unbase64}
\alias{unbase64}
\alias{unbase64}
\alias{unbase64,Column-method}
\title{unbase64}
\usage{
unbase64(x)

\S4method{unbase64}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Decodes a BASE64 encoded string column and returns it as a binary column.
This is the reverse of base64.
}
\note{
unbase64 since 1.5.0
}
\examples{
\dontrun{unbase64(df$c)}
}
\seealso{
Other string_funcs: \code{\link{ascii}},
  \code{\link{base64}}, \code{\link{concat_ws}},
  \code{\link{concat}}, \code{\link{decode}},
  \code{\link{encode}}, \code{\link{format_number}},
  \code{\link{format_string}}, \code{\link{initcap}},
  \code{\link{instr}}, \code{\link{length}},
  \code{\link{levenshtein}}, \code{\link{locate}},
  \code{\link{lower}}, \code{\link{lpad}},
  \code{\link{ltrim}}, \code{\link{regexp_extract}},
  \code{\link{regexp_replace}}, \code{\link{reverse}},
  \code{\link{rpad}}, \code{\link{rtrim}},
  \code{\link{soundex}}, \code{\link{substring_index}},
  \code{\link{translate}}, \code{\link{trim}},
  \code{\link{upper}}
}
