% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{uncacheTable}
\alias{uncacheTable}
\title{Uncache Table}
\usage{
uncacheTable(sqlContext, tableName)
}
\arguments{
\item{sqlContext}{SQLContext to use}

\item{tableName}{The name of the table being uncached}
}
\value{
DataFrame
}
\description{
Removes the specified table from the in-memory cache.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
registerTempTable(df, "table")
uncacheTable(sqlContext, "table")
}
}

