% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{unionAll,DataFrame,DataFrame-method}
\alias{unionAll}
\alias{unionAll,DataFrame,DataFrame-method}
\title{UnionAll}
\usage{
\S4method{unionAll}{DataFrame,DataFrame}(x, y)

unionAll(x, y)
}
\arguments{
\item{x}{A Spark DataFrame}

\item{y}{A Spark DataFrame}
}
\value{
A DataFrame containing the result of the union.
}
\description{
Return a new DataFrame containing the union of rows in this DataFrame
and another DataFrame. This is equivalent to `UNION ALL` in SQL.
Note that this does not remove duplicate rows across the two DataFrames.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
df1 <- jsonFile(sqlContext, path)
df2 <- jsonFile(sqlContext, path2)
unioned <- unionAll(df, df2)
}
}

