% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{unix_timestamp}
\alias{unix_timestamp}
\alias{unix_timestamp}
\alias{unix_timestamp,missing,missing-method}
\alias{unix_timestamp}
\alias{unix_timestamp,Column,missing-method}
\alias{unix_timestamp}
\alias{unix_timestamp,Column,character-method}
\title{unix_timestamp}
\usage{
unix_timestamp(x, format)

\S4method{unix_timestamp}{missing,missing}(x, format)

\S4method{unix_timestamp}{Column,missing}(x, format)

\S4method{unix_timestamp}{Column,character}(x, format = "yyyy-MM-dd HH:mm:ss")
}
\arguments{
\item{x}{a Column of date, in string, date or timestamp type.}

\item{format}{the target format. See
\href{http://docs.oracle.com/javase/tutorial/i18n/format/simpleDateFormat.html}{
Customizing Formats} for available options.}
}
\description{
Gets current Unix timestamp in seconds.
}
\note{
unix_timestamp since 1.5.0

unix_timestamp(Column) since 1.5.0

unix_timestamp(Column, character) since 1.5.0
}
\examples{
\dontrun{
unix_timestamp()
unix_timestamp(df$t)
unix_timestamp(df$t, 'yyyy-MM-dd HH')
}
}
\seealso{
Other datetime_funcs: \code{\link{add_months}},
  \code{\link{date_add}}, \code{\link{date_format}},
  \code{\link{date_sub}}, \code{\link{datediff}},
  \code{\link{dayofmonth}}, \code{\link{dayofyear}},
  \code{\link{from_unixtime}},
  \code{\link{from_utc_timestamp}}, \code{\link{hour}},
  \code{\link{last_day}}, \code{\link{minute}},
  \code{\link{months_between}}, \code{\link{month}},
  \code{\link{next_day}}, \code{\link{quarter}},
  \code{\link{second}}, \code{\link{to_date}},
  \code{\link{to_utc_timestamp}}, \code{\link{weekofyear}},
  \code{\link{window}}, \code{\link{year}}
}
