% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R
\docType{methods}
\name{unpersist,DataFrame-method}
\alias{unpersist,DataFrame-method}
\title{Unpersist}
\usage{
\S4method{unpersist}{DataFrame}(x, blocking = TRUE)
}
\arguments{
\item{x}{The DataFrame to unpersist}

\item{blocking}{Whether to block until all blocks are deleted}
}
\description{
Mark this DataFrame as non-persistent, and remove all blocks for it from memory and
disk.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
persist(df, "MEMORY_AND_DISK")
unpersist(df)
}
}

