% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{var}
\alias{var}
\alias{variance}
\alias{var}
\alias{var,Column-method}
\alias{variance}
\alias{variance,Column-method}
\title{var}
\usage{
var(x, y = NULL, na.rm = FALSE, use)

variance(x)

\S4method{var}{Column}(x)

\S4method{variance}{Column}(x)
}
\arguments{
\item{x}{a Column to compute on.}

\item{y, na.rm, use}{currently not used.}
}
\description{
Aggregate function: alias for \link{var_samp}.
}
\note{
var since 1.6.0

variance since 1.6.0
}
\examples{
\dontrun{
variance(df$c)
select(df, var_pop(df$age))
agg(df, var(df$age))
}
}
\seealso{
\link{var_pop}, \link{var_samp}

Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{countDistinct}}, \code{\link{count}},
  \code{\link{first}}, \code{\link{kurtosis}},
  \code{\link{last}}, \code{\link{max}},
  \code{\link{mean}}, \code{\link{min}}, \code{\link{sd}},
  \code{\link{skewness}}, \code{\link{stddev_pop}},
  \code{\link{stddev_samp}}, \code{\link{sumDistinct}},
  \code{\link{sum}}, \code{\link{var_pop}},
  \code{\link{var_samp}}
}
