% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{var_pop}
\alias{var_pop}
\alias{var_pop}
\alias{var_pop,Column-method}
\title{var_pop}
\usage{
var_pop(x)

\S4method{var_pop}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Aggregate function: returns the population variance of the values in a group.
}
\note{
var_pop since 1.5.0
}
\examples{
\dontrun{var_pop(df$c)}
}
\seealso{
\link{var}, \link{var_samp}

Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{countDistinct}}, \code{\link{count}},
  \code{\link{first}}, \code{\link{kurtosis}},
  \code{\link{last}}, \code{\link{max}},
  \code{\link{mean}}, \code{\link{min}}, \code{\link{sd}},
  \code{\link{skewness}}, \code{\link{stddev_pop}},
  \code{\link{stddev_samp}}, \code{\link{sumDistinct}},
  \code{\link{sum}}, \code{\link{var_samp}},
  \code{\link{var}}
}
