% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{var_samp}
\alias{var_samp}
\alias{var_samp}
\alias{var_samp,Column-method}
\title{var_samp}
\usage{
var_samp(x)

\S4method{var_samp}{Column}(x)
}
\arguments{
\item{x}{Column to compute on.}
}
\description{
Aggregate function: returns the unbiased variance of the values in a group.
}
\note{
var_samp since 1.6.0
}
\examples{
\dontrun{var_samp(df$c)}
}
\seealso{
\link{var_pop}, \link{var}

Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{countDistinct}}, \code{\link{count}},
  \code{\link{first}}, \code{\link{kurtosis}},
  \code{\link{last}}, \code{\link{max}},
  \code{\link{mean}}, \code{\link{min}}, \code{\link{sd}},
  \code{\link{skewness}}, \code{\link{stddev_pop}},
  \code{\link{stddev_samp}}, \code{\link{sumDistinct}},
  \code{\link{sum}}, \code{\link{var_pop}},
  \code{\link{var}}
}
