% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/functions.R
\docType{methods}
\name{when}
\alias{when}
\alias{when}
\alias{when,Column-method}
\title{when}
\usage{
when(condition, value)

\S4method{when}{Column}(condition, value)
}
\arguments{
\item{condition}{the condition to test on. Must be a Column expression.}

\item{value}{result expression.}
}
\description{
Evaluates a list of conditions and returns one of multiple possible result expressions.
For unmatched expressions null is returned.
}
\note{
when since 1.5.0
}
\examples{
\dontrun{when(df$age == 2, df$age + 1)}
}
\seealso{
\link{ifelse}

Other normal_funcs: \code{\link{abs}},
  \code{\link{bitwiseNOT}}, \code{\link{coalesce}},
  \code{\link{column}}, \code{\link{expr}},
  \code{\link{greatest}}, \code{\link{ifelse}},
  \code{\link{isnan}}, \code{\link{least}},
  \code{\link{lit}}, \code{\link{nanvl}},
  \code{\link{negate}}, \code{\link{randn}},
  \code{\link{rand}}, \code{\link{struct}}
}
