% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/DataFrame.R, R/generics.R
\docType{methods}
\name{withColumnRenamed,DataFrame,character,character-method}
\alias{rename}
\alias{rename,DataFrame-method}
\alias{withColumnRenamed}
\alias{withColumnRenamed,DataFrame,character,character-method}
\title{WithColumnRenamed}
\usage{
\S4method{withColumnRenamed}{DataFrame,character,character}(x, existingCol,
  newCol)

\S4method{rename}{DataFrame}(x, ...)

rename(x, ...)

withColumnRenamed(x, existingCol, newCol)
}
\arguments{
\item{x}{A DataFrame}

\item{existingCol}{The name of the column you want to change.}

\item{newCol}{The new column name.}

\item{x}{A DataFrame}

\item{newCol}{A named pair of the form new_column_name = existing_column}
}
\value{
A DataFrame with the column name changed.

A DataFrame with the column name changed.
}
\description{
Rename an existing column in a DataFrame.

Rename an existing column in a DataFrame.
}
\examples{
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
newDF <- withColumnRenamed(df, "col1", "newCol1")
}
\dontrun{
sc <- sparkR.init()
sqlContext <- sparkRSQL.init(sc)
path <- "path/to/file.json"
df <- jsonFile(sqlContext, path)
newDF <- rename(df, col1 = df$newCol1)
}
}

