% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/mllib.R
\name{write.ml}
\alias{write.ml}
\alias{write.ml}
\title{Saves the MLlib model to the input path}
\usage{
write.ml(object, path, ...)
}
\arguments{
\item{object}{a fitted ML model object.}

\item{path}{the directory where the model is saved.}

\item{...}{additional argument(s) passed to the method.}
}
\description{
Saves the MLlib model to the input path. For more information, see the specific
MLlib model below.
}
\seealso{
\link{spark.glm}, \link{glm},

\link{spark.als}, \link{spark.gaussianMixture}, \link{spark.gbt}, \link{spark.isoreg},

\link{spark.kmeans},

\link{spark.lda}, \link{spark.logit}, \link{spark.mlp}, \link{spark.naiveBayes},

\link{spark.randomForest}, \link{spark.survreg},

\link{read.ml}
}
