% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{cacheTable}
\alias{cacheTable}
\title{Cache Table}
\usage{
cacheTable(tableName)
}
\arguments{
\item{tableName}{the qualified or unqualified name that designates a table. If no database
identifier is provided, it refers to a table in the current database.}
}
\value{
SparkDataFrame
}
\description{
Caches the specified table in-memory.
}
\note{
cacheTable since 1.4.0
}
\examples{
\dontrun{
sparkR.session()
path <- "path/to/file.json"
df <- read.json(path)
createOrReplaceTempView(df, "table")
cacheTable("table")
}
}
