% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{cacheTable}
\alias{cacheTable}
\alias{cacheTable.default}
\title{Cache Table}
\usage{
\method{cacheTable}{default}(tableName)
}
\arguments{
\item{tableName}{The name of the table being cached}
}
\value{
SparkDataFrame
}
\description{
Caches the specified table in-memory.
}
\note{
cacheTable since 1.4.0
}
\examples{
\dontrun{
sparkR.session()
path <- "path/to/file.json"
df <- read.json(path)
createOrReplaceTempView(df, "table")
cacheTable("table")
}
}
