% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{clearCache}
\alias{clearCache}
\alias{clearCache.default}
\title{Clear Cache}
\usage{
\method{clearCache}{default}()
}
\description{
Removes all cached tables from the in-memory cache.
}
\note{
clearCache since 1.4.0
}
\examples{
\dontrun{
clearCache()
}
}
