% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/group.R, R/functions.R
\docType{methods}
\name{count}
\alias{count}
\alias{n}
\alias{count,GroupedData-method}
\alias{count}
\alias{count,Column-method}
\alias{n}
\alias{n,Column-method}
\title{Count}
\usage{
count(x)

n(x)

\S4method{count}{GroupedData}(x)

\S4method{count}{Column}(x)

\S4method{n}{Column}(x)
}
\arguments{
\item{x}{a GroupedData or Column.}
}
\value{
A SparkDataFrame.
}
\description{
Count the number of rows for each group when we have \code{GroupedData} input.
The resulting SparkDataFrame will also contain the grouping columns.

This can be used as a column aggregate function with \code{Column} as input,
and returns the number of items in a group.
}
\note{
count since 1.4.0

count since 1.4.0

n since 1.4.0
}
\examples{
\dontrun{
  count(groupBy(df, "name"))
}
\dontrun{count(df$c)}
\dontrun{n(df$c)}
}
\seealso{
Other agg_funcs: \code{\link{agg}}, \code{\link{avg}},
  \code{\link{countDistinct}}, \code{\link{first}},
  \code{\link{kurtosis}}, \code{\link{last}},
  \code{\link{max}}, \code{\link{mean}}, \code{\link{min}},
  \code{\link{sd}}, \code{\link{skewness}},
  \code{\link{stddev_pop}}, \code{\link{stddev_samp}},
  \code{\link{sumDistinct}}, \code{\link{sum}},
  \code{\link{var_pop}}, \code{\link{var_samp}},
  \code{\link{var}}
}
