% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/group.R, R/functions.R
\name{count}
\alias{count}
\alias{n}
\alias{count,GroupedData-method}
\alias{count,Column-method}
\alias{n,Column-method}
\title{Count}
\usage{
count(x)

n(x)

\S4method{count}{GroupedData}(x)

\S4method{count}{Column}(x)

\S4method{n}{Column}(x)
}
\arguments{
\item{x}{a GroupedData or Column.}
}
\value{
A SparkDataFrame.
}
\description{
Count the number of rows for each group when we have \code{GroupedData} input.
The resulting SparkDataFrame will also contain the grouping columns.

This can be used as a column aggregate function with \code{Column} as input,
and returns the number of items in a group.
}
\note{
count since 1.4.0

count since 1.4.0

n since 1.4.0
}
\examples{
\dontrun{
  count(groupBy(df, "name"))
}
\dontrun{count(df$c)}
\dontrun{n(df$c)}
}
\seealso{
Other aggregate functions: 
\code{\link{avg}()},
\code{\link{column_aggregate_functions}},
\code{\link{corr}()},
\code{\link{cov}()},
\code{\link{first}()},
\code{\link{last}()}
}
\concept{aggregate functions}
