% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.R
\docType{class}
\name{GroupedData-class}
\alias{GroupedData-class}
\alias{groupedData}
\title{S4 class that represents a GroupedData}
\usage{
groupedData(sgd)
}
\arguments{
\item{sgd}{A Java object reference to the backing Scala GroupedData}
}
\description{
GroupedDatas can be created using groupBy() on a SparkDataFrame
}
\note{
GroupedData since 1.4.0
}
\seealso{
groupBy
}
