% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{read.parquet}
\alias{read.parquet}
\alias{read.parquet.default}
\alias{parquetFile}
\alias{parquetFile.default}
\title{Create a SparkDataFrame from a Parquet file.}
\usage{
\method{read.parquet}{default}(path, ...)

\method{parquetFile}{default}(...)
}
\arguments{
\item{path}{path of file to read. A vector of multiple paths is allowed.}

\item{...}{argument(s) passed to the method.}
}
\value{
SparkDataFrame
}
\description{
Loads a Parquet file, returning the result as a SparkDataFrame.
}
\note{
read.parquet since 1.6.0

parquetFile since 1.4.0
}
