% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/DataFrame.R
\name{registerTempTable}
\alias{registerTempTable}
\alias{registerTempTable,SparkDataFrame,character-method}
\title{(Deprecated) Register Temporary Table}
\usage{
registerTempTable(x, tableName)

\S4method{registerTempTable}{SparkDataFrame,character}(x, tableName)
}
\arguments{
\item{x}{A SparkDataFrame}

\item{tableName}{A character vector containing the name of the table}
}
\description{
Registers a SparkDataFrame as a Temporary Table in the SparkSession
}
\note{
registerTempTable since 1.4.0
}
\examples{
\dontrun{
sparkR.session()
path <- "path/to/file.json"
df <- read.json(path)
registerTempTable(df, "json_df")
new_df <- sql("SELECT * FROM json_df")
}
}
\seealso{
\link{createOrReplaceTempView}
}
