% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/DataFrame.R
\docType{methods}
\name{sample}
\alias{sample}
\alias{sample_frac}
\alias{sample}
\alias{sample,SparkDataFrame,logical,numeric-method}
\alias{sample_frac}
\alias{sample_frac,SparkDataFrame,logical,numeric-method}
\title{Sample}
\usage{
sample(x, withReplacement, fraction, seed)

sample_frac(x, withReplacement, fraction, seed)

\S4method{sample}{SparkDataFrame,logical,numeric}(x, withReplacement, fraction,
  seed)

\S4method{sample_frac}{SparkDataFrame,logical,numeric}(x, withReplacement,
  fraction, seed)
}
\arguments{
\item{x}{A SparkDataFrame}

\item{withReplacement}{Sampling with replacement or not}

\item{fraction}{The (rough) sample target fraction}

\item{seed}{Randomness seed value}
}
\description{
Return a sampled subset of this SparkDataFrame using a random seed.
Note: this is not guaranteed to provide exactly the fraction specified
of the total count of of the given SparkDataFrame.
}
\note{
sample since 1.4.0

sample_frac since 1.4.0
}
\examples{
\dontrun{
sparkR.session()
path <- "path/to/file.json"
df <- read.json(path)
collect(sample(df, FALSE, 0.5))
collect(sample(df, TRUE, 0.5))
}
}
\seealso{
Other SparkDataFrame functions: \code{\link{SparkDataFrame-class}},
  \code{\link{agg}}, \code{\link{arrange}},
  \code{\link{as.data.frame}}, \code{\link{attach}},
  \code{\link{cache}}, \code{\link{coalesce}},
  \code{\link{collect}}, \code{\link{colnames}},
  \code{\link{coltypes}},
  \code{\link{createOrReplaceTempView}},
  \code{\link{crossJoin}}, \code{\link{dapplyCollect}},
  \code{\link{dapply}}, \code{\link{describe}},
  \code{\link{dim}}, \code{\link{distinct}},
  \code{\link{dropDuplicates}}, \code{\link{dropna}},
  \code{\link{drop}}, \code{\link{dtypes}},
  \code{\link{except}}, \code{\link{explain}},
  \code{\link{filter}}, \code{\link{first}},
  \code{\link{gapplyCollect}}, \code{\link{gapply}},
  \code{\link{getNumPartitions}}, \code{\link{group_by}},
  \code{\link{head}}, \code{\link{histogram}},
  \code{\link{insertInto}}, \code{\link{intersect}},
  \code{\link{isLocal}}, \code{\link{join}},
  \code{\link{limit}}, \code{\link{merge}},
  \code{\link{mutate}}, \code{\link{ncol}},
  \code{\link{nrow}}, \code{\link{persist}},
  \code{\link{printSchema}}, \code{\link{randomSplit}},
  \code{\link{rbind}}, \code{\link{registerTempTable}},
  \code{\link{rename}}, \code{\link{repartition}},
  \code{\link{saveAsTable}}, \code{\link{schema}},
  \code{\link{selectExpr}}, \code{\link{select}},
  \code{\link{showDF}}, \code{\link{show}},
  \code{\link{storageLevel}}, \code{\link{str}},
  \code{\link{subset}}, \code{\link{take}},
  \code{\link{union}}, \code{\link{unpersist}},
  \code{\link{withColumn}}, \code{\link{with}},
  \code{\link{write.df}}, \code{\link{write.jdbc}},
  \code{\link{write.json}}, \code{\link{write.orc}},
  \code{\link{write.parquet}}, \code{\link{write.text}}
}
