% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparkr.R
\name{sparkr}
\alias{sparkr}
\title{Give a random idea or "spark" of an activity to do.}
\usage{
sparkr(give_idea = TRUE)
}
\arguments{
\item{give_idea}{Logical. Set to TRUE to get a randomly selected idea of something to do}
}
\value{
Gives the product of \code{give_idea}. A random idea from a predefined list of activities.
}
\description{
Give a random idea or "spark" of an activity to do.
}
\examples{
sparkr(TRUE)
sparkr(give_idea = TRUE)
sparkr(give_idea = FALSE)

}
