% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{sql}
\alias{sql}
\title{SQL Query}
\usage{
sql(sqlQuery)
}
\arguments{
\item{sqlQuery}{A character vector containing the SQL query}
}
\value{
SparkDataFrame
}
\description{
Executes a SQL query using Spark, returning the result as a SparkDataFrame.
}
\note{
sql since 1.4.0
}
\examples{
\dontrun{
sparkR.session()
path <- "path/to/file.json"
df <- read.json(path)
createOrReplaceTempView(df, "table")
new_df <- sql("SELECT * FROM table")
}
}
