% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{substr}
\alias{substr}
\alias{substr,Column-method}
\title{substr}
\usage{
\S4method{substr}{Column}(x, start, stop)
}
\arguments{
\item{x}{a Column.}

\item{start}{starting position. It should be 1-base.}

\item{stop}{ending position.}
}
\description{
An expression that returns a substring.
}
\note{
substr since 1.4.0
}
\examples{
\dontrun{
df <- createDataFrame(list(list(a="abcdef")))
collect(select(df, substr(df$a, 1, 4))) # the result is `abcd`.
collect(select(df, substr(df$a, 2, 4))) # the result is `bcd`.
}
}
\seealso{
Other column_func: 
\code{\link{alias}()},
\code{\link{between}()},
\code{\link{cast}()},
\code{\link{endsWith}()},
\code{\link{otherwise}()},
\code{\link{over}()},
\code{\link{startsWith}()}
}
\concept{column_func}
