% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{tableNames}
\alias{tableNames}
\alias{tableNames.default}
\title{Table Names}
\usage{
\method{tableNames}{default}(databaseName = NULL)
}
\arguments{
\item{databaseName}{name of the database}
}
\value{
a list of table names
}
\description{
Returns the names of tables in the given database as an array.
}
\note{
tableNames since 1.4.0
}
\examples{
\dontrun{
sparkR.session()
tableNames("hive")
}
}
