% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{tables}
\alias{tables}
\alias{tables.default}
\title{Tables}
\usage{
\method{tables}{default}(databaseName = NULL)
}
\arguments{
\item{databaseName}{name of the database}
}
\value{
a SparkDataFrame
}
\description{
Returns a SparkDataFrame containing names of tables in the given database.
}
\note{
tables since 1.4.0
}
\examples{
\dontrun{
sparkR.session()
tables("hive")
}
}
