% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQLContext.R
\name{uncacheTable}
\alias{uncacheTable}
\alias{uncacheTable.default}
\title{Uncache Table}
\usage{
\method{uncacheTable}{default}(tableName)
}
\arguments{
\item{tableName}{The name of the table being uncached}
}
\value{
SparkDataFrame
}
\description{
Removes the specified table from the in-memory cache.
}
\note{
uncacheTable since 1.4.0
}
\examples{
\dontrun{
sparkR.session()
path <- "path/to/file.json"
df <- read.json(path)
createOrReplaceTempView(df, "table")
uncacheTable("table")
}
}
